+++ Testes/server/.env
@@ -1,0 +1,8 @@
+PORT = 3011
+JWT_SECRET = @tokenJWT
+
+DB_USER = postgres
+DB_HOST = localhost
+DB_NAME = bd_nutri
+DB_PASSWORD = 2e6i7ef2r
+DB_PORT = 5432
