+++ Testes/server/src/controllers/index.ts
@@ -1,0 +1,6 @@
+
+import UserController from "./UserController";
+
+export {
+    UserController,
+};
