+++ Testes/server/src/database/load.ts
@@ -1,0 +1,644 @@
+import {pool, queryCommitRollback} from "../database/connection";
+
+async function load() {
+  try {
+    await queryCommitRollback(`
+    DO
+    $$
+    BEGIN
+
+        TRUNCATE foods CASCADE;
+        TRUNCATE categories CASCADE;
+
+
+        INSERT INTO categories (id,name)
+        VALUES 
+        (1,'Cereais e derivados'),
+        (2,'Verduras, hortaliças e derivados'),
+        (3,'Gorduras e óleos'),
+        (4,'Frutas e derivados'),
+        (5,'Pescados e frutos do mar'),
+        (6,'Carnes e derivados'),
+        (7,'Leite e derivados'),
+        (8,'Bebidas (alcoólicas e não alcoólicas)'),
+        (9,'Ovos e derivados'),
+        (10,'Produtos açucarados'),
+        (11,'Miscelâneas'),
+        (12,'Outros alimentos industrializados'),
+        (13,'Alimentos preparados'),
+        (14,'Leguminosas e derivados'),
+        (15,'Nozes e sementes');
+
+        INSERT INTO foods (id,category,description,moisture,energy,protein,lipids,cholesterol,carbohydrate,dietary_fiber,ash,calcium,magnesium,manganese,phosphorus,iron,sodium,potassium,copper,zinc,retinol,re,era,thiamin,riboflavin,pyridoxine,niacin,vitamin_c)
+        VALUES
+       (1,1,'Arroz, integral, cozido',70.1,124,2.6,1,NULL,25.8,2.7,0.5,5,59,0.63,106,0.3,1,75,0.02,0.7,NULL,NULL,NULL,0.08,NULL,0.08,NULL,NULL),
+(2,1,'Arroz, integral, cru',12.2,360,7.3,1.9,NULL,77.5,4.8,1.2,8,110,2.99,251,0.9,2,173,0.07,1.4,NULL,NULL,NULL,0.26,NULL,0.18,4.18,NULL),
+(3,1,'Arroz, tipo 1, cozido',69.1,128,2.5,0.2,NULL,28.1,1.6,0.1,4,2,0.3,18,0.1,1,15,0.02,0.5,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(4,1,'Arroz, tipo 1, cru',13.2,358,7.2,0.3,NULL,78.8,1.6,0.5,4,30,1.03,104,0.7,1,62,0.11,1.2,NULL,NULL,NULL,0.16,NULL,0.07,1.12,NULL),
+(5,1,'Arroz, tipo 2, cozido',68.7,130,2.6,0.4,NULL,28.2,1.1,0.1,3,6,0.37,22,0.1,2,20,0.04,0.5,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(6,1,'Arroz, tipo 2, cru',13.2,358,7.2,0.3,NULL,78.9,1.7,0.4,5,29,0.83,82,0.6,1,57,0.05,1.3,NULL,NULL,NULL,0.16,NULL,0.05,0.92,NULL),
+(7,1,'Aveia, flocos, crua',9.1,394,13.9,8.5,NULL,66.6,9.1,1.8,48,119,1.89,153,4.4,5,336,0.44,2.6,NULL,NULL,NULL,0.55,0.03,NULL,4.47,1.4),
+(8,1,'Biscoito, doce, maisena',3.2,443,8.1,12,NULL,75.2,2.1,1.5,54,37,0.78,166,1.8,352,142,0.17,1,NULL,NULL,NULL,1.01,0.42,0.23,3.91,6.2),
+(9,1,'Biscoito, doce, recheado com chocolate',2.2,472,6.4,19.6,NULL,70.5,3,1.3,27,48,0.59,139,2.3,239,232,0.27,1,NULL,NULL,NULL,0.32,0.39,0.46,2.52,3.5),
+(10,1,'Biscoito, doce, recheado com morango',2.7,471,5.7,19.6,NULL,71,1.5,1,36,27,0.66,138,1.5,230,113,0.13,0.7,NULL,NULL,NULL,0.9,0.42,0.21,1.5,NULL),
+(11,1,'Biscoito, doce, wafer, recheado de chocolate',1.2,502,5.6,24.7,NULL,67.5,1.8,1.1,23,48,0.44,124,2.4,137,240,0.26,0.9,NULL,NULL,NULL,0.37,0.03,0.3,1.12,NULL),
+(12,1,'Biscoito, doce, wafer, recheado de morango',1.2,513,4.5,26.4,1,67.4,0.8,0.6,14,19,0.28,73,1.1,120,75,0.08,0.5,NULL,NULL,NULL,0.36,NULL,0.94,0.37,NULL),
+(13,1,'Biscoito, salgado, cream cracker',4.1,432,10.1,14.4,NULL,68.7,2.5,2.7,20,40,0.69,148,2.2,854,181,0.18,1.1,NULL,NULL,NULL,0.71,0.13,0.17,7.14,NULL),
+(14,1,'Bolo, mistura para',1,419,6.2,6.1,NULL,84.7,1.7,2,59,28,0.46,333,1.2,463,75,0.15,0.6,NULL,NULL,NULL,0.18,NULL,1.5,1.52,NULL),
+(15,1,'Bolo, pronto, aipim',34.1,324,4.4,12.7,73,47.9,0.7,0.8,85,10,0.11,122,0.5,111,135,0.05,0.4,NULL,NULL,NULL,0.13,0.07,NULL,NULL,NULL),
+(16,1,'Bolo, pronto, chocolate',19.3,410,6.2,18.5,77,54.7,1.4,1.3,75,28,0.38,197,2.1,283,212,0.05,0.7,NULL,NULL,NULL,0.13,0.09,0.05,NULL,NULL),
+(17,1,'Bolo, pronto, coco',29.3,333,5.7,11.3,63,52.3,1.1,1.4,57,16,0.4,303,0.8,190,143,0.09,0.7,NULL,NULL,NULL,0.15,0.06,NULL,NULL,NULL),
+(18,1,'Bolo, pronto, milho',36.7,311,4.8,12.4,82,45.1,0.7,1,83,10,0.11,128,0.7,134,118,0.04,0.4,57,NULL,NULL,0.11,0.05,NULL,NULL,NULL),
+(19,1,'Canjica, branca, crua',13.6,358,7.2,1,NULL,78.1,5.5,0.2,2,12,0.09,48,0.3,1,93,0.05,0.4,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(20,1,'Canjica, com leite integral',72.5,112,2.4,1.2,1,23.6,1.2,0.3,43,6,0.02,41,0.1,28,70,0.03,0.3,NULL,NULL,NULL,NULL,0.04,NULL,NULL,NULL),
+(21,1,'Cereais, milho, flocos, com sal',9.3,370,7.3,1.6,NULL,80.8,5.3,1,2,20,NULL,91,0.5,272,69,NULL,0.6,NULL,NULL,NULL,0.12,NULL,0.06,NULL,NULL),
+(22,1,'Cereais, milho, flocos, sem sal',11.2,363,6.9,1.2,NULL,80.4,1.8,0.3,2,17,NULL,58,1.7,31,29,0.24,0.3,NULL,NULL,NULL,0.05,NULL,0.04,NULL,NULL),
+(23,1,'Cereais, mingau, milho, infantil',4.7,394,6.4,1.1,NULL,87.3,3.2,0.5,219,16,0.11,169,3,399,82,0.04,0.4,21,NULL,NULL,3.72,0.47,5.08,24.16,109.4),
+(24,1,'Cereais, mistura para vitamina, trigo, cevada e aveia',4.4,381,8.9,2.1,NULL,81.6,5,3,584,72,2.28,515,12.6,1163,244,0.21,2,NULL,NULL,NULL,0.76,0.88,1.49,7.46,13.1),
+(25,1,'Cereal matinal, milho',5.5,365,7.2,1,NULL,83.8,4.1,2.5,143,11,0.06,101,3.1,655,83,0.06,7.6,NULL,31,15,0.76,1.02,2.25,11.04,17.3),
+(26,1,'Cereal matinal, milho, açúcar',4.3,377,4.7,0.7,NULL,88.8,2.1,1.5,56,8,0.13,43,3.9,405,52,0.04,8.5,NULL,31,15,0.73,1.11,0.79,10.13,14.6),
+(27,1,'Creme de arroz, pó',7.3,386,7,1.2,NULL,83.9,1.1,0.5,7,51,1.24,153,0.6,1,115,0.04,1.9,NULL,NULL,NULL,0.22,NULL,0.05,NULL,NULL),
+(28,1,'Creme de milho, pó',5.7,333,4.8,1.6,NULL,86.1,3.7,1.7,323,30,0.2,303,4.3,594,166,0.04,0.8,NULL,NULL,NULL,0.74,NULL,2.05,NULL,96.3),
+(29,1,'Curau, milho verde',81.6,78,2.4,1.6,5,13.9,0.5,0.5,53,16,0.06,75,0.4,21,162,0.03,0.4,12,20,16,0.04,0.07,NULL,NULL,NULL),
+(30,1,'Curau, milho verde, mistura para',3.9,402,2.2,13.4,NULL,79.8,2.5,0.7,31,9,0.05,45,0.9,223,55,0.03,0.2,NULL,NULL,NULL,0.13,NULL,NULL,NULL,NULL),
+(31,1,'Farinha, de arroz, enriquecida',12.7,363,1.3,0.3,NULL,85.5,0.6,0.2,1,4,0.04,36,31.4,17,13,NULL,8.5,NULL,NULL,NULL,3.23,NULL,3.47,24.42,173.6),
+(32,1,'Farinha, de centeio, integral',10.8,336,12.5,1.8,NULL,73.3,15.5,1.7,34,120,3.86,340,4.7,41,334,0.56,2.7,NULL,NULL,NULL,0.29,0.03,0.08,NULL,NULL),
+(33,1,'Farinha, de milho, amarela',11.8,351,7.2,1.5,NULL,79.1,5.5,0.5,1,31,NULL,84,2.3,45,58,0.27,0.6,NULL,18,9,0.25,NULL,0.25,NULL,NULL),
+(34,1,'Farinha, de rosca',9.8,371,11.4,1.5,NULL,75.8,4.8,1.6,35,57,1.62,195,6.7,333,212,0.26,1.7,NULL,NULL,NULL,0.25,NULL,0.09,NULL,NULL),
+(35,1,'Farinha, de trigo',13,360,9.8,1.4,NULL,75.1,2.3,0.8,18,31,0.46,115,1,1,151,0.15,0.8,NULL,NULL,NULL,0.31,NULL,NULL,0.89,NULL),
+(36,1,'Farinha, láctea, de cereais',2.7,415,11.9,5.8,11,77.8,1.9,1.9,196,58,1.49,296,8.7,125,366,0.19,1.7,492,NULL,NULL,1.43,1.13,1.14,9.5,24.3),
+(37,1,'Lasanha, massa fresca, cozida',59.6,164,5.8,1.2,NULL,32.5,1.6,0.9,10,4,0.2,42,1.2,207,54,0.07,0.4,NULL,NULL,NULL,0.04,NULL,NULL,3.96,NULL),
+(38,1,'Lasanha, massa fresca, crua',45,220,7,1.3,NULL,45.1,1.6,1.6,17,13,0.34,82,1.9,667,137,0.1,0.8,NULL,NULL,NULL,0.08,NULL,NULL,NULL,NULL),
+(39,1,'Macarrão, instantâneo',6,436,8.8,17.2,NULL,62.4,5.6,5.6,18,19,0.25,112,0.8,1516,148,0.1,0.5,NULL,NULL,NULL,1.18,0.04,0.53,9.37,NULL),
+(40,1,'Macarrão, trigo, cru',10.2,371,10,1.3,NULL,77.9,2.9,0.5,17,28,0.53,100,0.9,7,147,0.15,0.8,NULL,NULL,NULL,0.18,0.02,NULL,3.57,NULL),
+(41,1,'Macarrão, trigo, cru, com ovos',10.6,371,10.3,2,18,76.6,2.3,0.5,19,NULL,0.4,118,0.9,15,134,0.14,0.8,NULL,NULL,NULL,0.11,0.05,0.03,4.37,NULL),
+(42,1,'Milho, amido, cru',12.2,361,0.6,NULL,NULL,87.1,0.7,0.1,1,3,0.02,13,0.1,8,9,0.02,0.1,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(43,1,'Milho, fubá, cru',11.5,353,7.2,1.9,NULL,78.9,4.7,0.6,3,41,0.34,108,0.9,NULL,168,0.08,1.1,NULL,26,13,0.25,NULL,NULL,0.75,NULL),
+(44,1,'Milho, verde, cru',63.5,138,6.6,0.6,NULL,28.6,3.9,0.7,2,33,0.12,113,0.4,1,185,0.05,0.5,NULL,32,16,0.3,NULL,0.04,NULL,NULL),
+(45,1,'Milho, verde, enlatado, drenado',76.2,98,3.2,2.4,NULL,17.1,4.6,1.1,2,20,0.09,61,0.6,260,162,0.05,0.5,NULL,48,24,NULL,0.05,NULL,3.74,1.7),
+(46,1,'Mingau tradicional, pó',8.1,373,0.6,0.4,NULL,89.3,0.9,1.6,522,4,NULL,273,42,15,NULL,NULL,15.2,1533,NULL,NULL,3.41,NULL,3.11,19.39,NULL),
+(47,1,'Pamonha, barra para cozimento, pré-cozida',61.3,171,2.6,4.8,NULL,30.7,2.4,0.6,4,15,0.12,55,0.4,132,125,0.03,0.4,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(48,1,'Pão, aveia, forma',19.9,343,12.4,5.7,NULL,59.6,6,2.5,109,57,1.08,182,4.7,606,210,0.14,1.7,NULL,NULL,NULL,0.09,0.03,0.08,NULL,NULL),
+(49,1,'Pão, de soja',26,309,11.3,3.6,NULL,56.5,5.7,2.5,90,48,0.57,153,3.3,663,296,0.16,1.5,NULL,NULL,NULL,0.07,0.04,0.79,NULL,NULL),
+(50,1,'Pão, glúten, forma',40.7,253,12,2.7,NULL,44.1,2.5,0.5,156,24,0.51,105,5.7,22,65,0.06,1.3,NULL,NULL,NULL,0.04,0.03,NULL,NULL,NULL),
+(51,1,'Pão, milho, forma',30.4,292,8.3,3.1,6,56.4,4.3,1.8,78,29,0.37,110,3,507,89,0.1,0.8,NULL,NULL,NULL,0.08,NULL,0.08,NULL,NULL),
+(52,1,'Pão, trigo, forma, integral',34.7,253,9.4,3.7,NULL,49.9,6.9,2.3,132,60,1.62,193,3,506,163,0.15,1.6,NULL,NULL,NULL,0.08,0.04,0.15,NULL,NULL),
+(53,1,'Pão, trigo, francês',28.5,300,8,3.1,NULL,58.6,2.3,1.8,16,25,0.46,95,1,648,142,0.13,0.8,3,NULL,NULL,0.39,0.67,0.6,2.34,NULL),
+(54,1,'Pão, trigo, sovado',25.8,311,8.4,2.8,17,61.5,2.4,1.5,52,22,0.31,101,2.3,431,91,NULL,2.7,NULL,NULL,NULL,0.08,0.04,0.15,NULL,NULL),
+(55,1,'Pastel, de carne, cru',34.4,289,10.7,8.8,18,42,1,4.1,17,18,0.34,117,2,1309,166,0.11,1.7,NULL,NULL,NULL,0.06,0.03,NULL,1.58,NULL),
+(56,1,'Pastel, de carne, frito',22.9,388,10.1,20.1,25,43.8,1,3.1,13,14,0.22,90,2.5,1040,156,0.12,1.2,NULL,NULL,NULL,0.06,0.05,NULL,1.83,NULL),
+(57,1,'Pastel, de queijo, cru',31.3,308,9.9,9.6,14,45.9,1.1,3.3,155,16,0.24,168,1,985,103,0.1,1,NULL,NULL,NULL,0.07,0.03,NULL,NULL,NULL),
+(58,1,'Pastel, de queijo, frito',17.5,422,8.7,22.7,15,48.1,0.9,2.9,126,15,0.2,124,1.3,821,124,0.1,0.8,18,NULL,NULL,0.09,0.04,NULL,NULL,NULL),
+(59,1,'Pastel, massa, crua',27.1,310,6.9,5.5,NULL,57.4,1.4,3.2,13,14,0.38,73,1.1,1344,167,0.11,0.6,NULL,NULL,NULL,0.15,NULL,NULL,NULL,NULL),
+(60,1,'Pastel, massa, frita',1,570,6,40.9,NULL,49.3,1.3,2.8,11,13,0.36,62,1.4,1175,143,0.08,0.5,NULL,NULL,NULL,0.16,NULL,NULL,NULL,NULL),
+(61,1,'Pipoca, com óleo de soja, sem sal',2.8,448,9.9,15.9,NULL,70.3,14.3,1,3,91,0.65,225,1.2,4,256,0.46,2,NULL,NULL,NULL,0.03,0.03,NULL,NULL,NULL),
+(62,1,'Polenta, pré-cozida',72.7,103,2.3,0.3,NULL,23.3,2.4,1.4,1,4,NULL,17,NULL,442,100,0.04,0.1,NULL,NULL,NULL,0.04,NULL,NULL,NULL,NULL),
+(63,1,'Torrada, pão francês',9,377,10.5,3.3,NULL,74.6,3.4,2.6,19,32,0.55,114,1.2,829,189,0.16,0.9,NULL,NULL,NULL,0.38,NULL,NULL,NULL,NULL),
+(64,2,'Abóbora, cabotian, cozida',86.4,48,1.4,0.7,NULL,10.8,2.5,0.7,8,9,0.26,33,0.3,1,199,0.06,0.3,NULL,NULL,NULL,0.08,NULL,0.07,NULL,7.5),
+(65,2,'Abóbora, cabotian, crua',88.5,39,1.7,0.5,NULL,8.4,2.2,0.8,18,9,0.11,26,0.4,NULL,351,0.06,0.3,NULL,446,223,NULL,NULL,0.1,NULL,5.1),
+(66,2,'Abóbora, menina brasileira, crua',95.7,14,0.6,NULL,NULL,3.3,1.2,0.4,9,4,0.01,12,0.2,NULL,165,0.02,NULL,NULL,1108,554,0.07,NULL,0.04,NULL,1.5),
+(67,2,'Abóbora, moranga, crua',95.9,12,1,0.1,NULL,2.7,1.7,0.4,3,2,0.01,8,NULL,NULL,125,0.05,0.1,NULL,278,139,NULL,NULL,0.06,NULL,9.6),
+(68,2,'Abóbora, moranga, refogada',92.5,29,0.4,0.8,NULL,6,1.5,0.4,19,7,0.02,12,0.1,3,183,0.04,0.1,NULL,95,48,0.05,NULL,0.05,NULL,6.7),
+(69,2,'Abóbora, pescoço, crua',92.5,24,0.7,0.1,NULL,6.1,2.3,0.6,9,7,0.07,32,0.3,1,264,0.09,0.2,NULL,NULL,NULL,NULL,NULL,0.06,NULL,2.1),
+(70,2,'Abobrinha, italiana, cozida',95.3,15,1.1,0.2,NULL,3,1.6,0.4,17,17,0.11,22,0.2,1,126,0.01,0.3,NULL,NULL,NULL,0.05,NULL,0.06,NULL,2.1),
+(71,2,'Abobrinha, italiana, crua',93.9,19,1.1,0.1,NULL,4.3,1.4,0.6,15,20,0.09,32,0.2,NULL,253,0.05,0.2,NULL,41,20,NULL,0.06,0.03,NULL,6.9),
+(72,2,'Abobrinha, italiana, refogada',93.5,24,1.1,0.8,NULL,4.2,1.4,0.4,21,13,0.14,32,0.4,2,194,0.02,0.3,NULL,42,21,0.04,NULL,NULL,NULL,7.5),
+(73,2,'Abobrinha, paulista, crua',90.9,31,0.6,0.1,NULL,7.9,2.6,0.5,19,9,0.11,33,0.2,1,213,0.1,0.2,NULL,NULL,NULL,NULL,NULL,0.03,NULL,17.5),
+(74,2,'Acelga, crua',93.2,21,1.4,0.1,NULL,4.6,1.1,0.6,43,10,0.11,40,0.3,1,240,0.1,0.3,NULL,NULL,NULL,0.04,NULL,0.16,NULL,22.6),
+(75,2,'Agrião, cru',93.9,17,2.7,0.2,NULL,2.3,2.1,0.9,133,18,0.28,51,3.1,7,218,0.1,0.7,NULL,458,229,0.11,0.23,0.09,1.19,60.1),
+(76,2,'Aipo, cru',93.8,19,0.8,0.1,NULL,4.3,1,1.1,65,9,0.18,28,0.7,10,274,0.31,0.1,NULL,916,458,NULL,NULL,0.18,NULL,5.9),
+(77,2,'Alface, americana, crua',97.2,9,0.6,0.1,NULL,1.7,1,0.3,14,6,0.12,19,0.3,7,136,0.02,0.2,NULL,NULL,NULL,0.03,NULL,0.04,NULL,11),
+(78,2,'Alface, crespa, crua',96.1,11,1.3,0.2,NULL,1.7,1.8,0.7,38,11,0.2,26,0.4,3,267,0.03,0.3,NULL,234,117,0.11,0.12,NULL,1.09,15.6),
+(79,2,'Alface, lisa, crua',95,14,1.7,0.1,NULL,2.4,2.3,0.8,28,9,0.33,26,0.6,4,349,0.03,0.3,NULL,368,184,0.09,0.08,0.07,0.75,21.4),
+(80,2,'Alface, roxa, crua',95.7,13,0.9,0.2,NULL,2.5,2,0.7,34,9,0.12,51,2.5,7,308,0.04,0.2,NULL,312,156,NULL,NULL,NULL,NULL,13.5),
+(81,2,'Alfavaca, crua',90.2,29,2.7,0.5,NULL,5.2,4.1,1.4,258,84,0.15,50,1.3,5,261,0.15,0.7,NULL,NULL,NULL,NULL,0.12,0.56,NULL,NULL),
+(82,2,'Alho, cru',67.5,113,7,0.2,NULL,23.9,4.3,1.3,14,21,0.24,149,0.8,5,535,0.15,0.8,NULL,NULL,NULL,0.18,NULL,0.44,NULL,NULL),
+(83,2,'Alho-poró, cru',91,32,1.4,0.1,NULL,6.9,2.5,0.6,34,11,0.1,36,0.6,2,224,0.29,0.2,NULL,16,8,0.06,NULL,0.08,0.34,14.1),
+(84,2,'Almeirão, cru',93.7,18,1.8,0.2,NULL,3.3,2.6,1,19,21,0.17,40,0.7,2,369,0.1,0.3,NULL,566,283,0.1,0.18,NULL,0.63,1.7),
+(85,2,'Almeirão, refogado',86.5,65,1.7,4.8,NULL,5.7,3.4,1.2,63,17,0.28,31,1.6,15,315,0.06,0.2,NULL,NULL,NULL,0.03,0.06,0.05,4.03,1.5),
+(86,2,'Batata, baroa, cozida',79.3,80,0.9,0.2,NULL,18.9,1.8,0.8,12,8,0.22,29,0.4,2,258,0.15,0.4,NULL,NULL,NULL,0.06,NULL,NULL,1.98,17.1),
+(87,2,'Batata, baroa, crua',73.7,101,1,0.2,NULL,24,2.1,1.1,17,12,0.07,45,0.3,NULL,505,0.05,0.2,NULL,25,12,0.05,NULL,0.12,NULL,7.6),
+(88,2,'Batata, doce, cozida',80.4,77,0.6,0.1,NULL,18.4,2.2,0.4,17,11,0.14,15,0.2,3,148,0.06,0.1,NULL,NULL,NULL,0.08,NULL,0.05,2.57,23.8),
+(89,2,'Batata, doce, crua',69.5,118,1.3,0.1,NULL,28.2,2.6,0.9,21,17,0.18,36,0.4,9,340,0.11,0.2,NULL,NULL,NULL,0.06,NULL,0.1,NULL,16.5),
+(90,2,'Batata, frita, tipo chips, industrializada',2.7,543,5.6,36.6,NULL,51.2,2.5,3.9,12,24,0.23,96,0.7,607,1014,0.12,0.6,NULL,NULL,NULL,0.2,NULL,0.13,2.61,NULL),
+(91,2,'Batata, inglesa, cozida',86.4,52,1.2,NULL,NULL,11.9,1.3,0.5,4,5,0.07,24,0.2,2,161,0.06,0.2,NULL,NULL,NULL,0.05,NULL,0.08,NULL,3.8),
+(92,2,'Batata, inglesa, crua',82.9,64,1.8,NULL,NULL,14.7,1.2,0.6,4,15,0.1,39,0.4,NULL,302,0.09,0.2,NULL,NULL,NULL,0.1,NULL,0.15,NULL,31.1),
+(93,2,'Batata, inglesa, frita',44.1,267,5,13.1,NULL,35.6,8.1,2.2,6,14,0.15,70,0.4,2,489,0.1,0.4,NULL,NULL,NULL,0.17,NULL,0.1,2.51,16.3),
+(94,2,'Batata, inglesa, sauté',83.1,68,1.3,0.9,NULL,14.1,1.4,0.6,4,6,0.08,32,0.3,8,199,0.05,0.2,NULL,NULL,NULL,0.07,NULL,0.09,1.38,NULL),
+(95,2,'Berinjela, cozida',94.4,19,0.7,0.1,NULL,4.5,2.5,0.3,11,9,0.11,15,0.2,1,105,0.04,0.1,NULL,NULL,NULL,0.04,NULL,NULL,NULL,NULL),
+(96,2,'Berinjela, crua',93.8,20,1.2,0.1,NULL,4.4,2.9,0.4,9,13,0.1,20,0.2,NULL,205,0.06,0.1,NULL,24,12,0.04,0.05,NULL,NULL,3),
+(97,2,'Beterraba, cozida',90.6,32,1.3,0.1,NULL,7.2,1.9,0.8,15,17,0.19,30,0.2,23,245,0.04,0.4,NULL,NULL,NULL,0.09,NULL,NULL,NULL,1.2),
+(98,2,'Beterraba, crua',86,49,1.9,0.1,NULL,11.1,3.4,0.9,18,24,1.23,19,0.3,10,375,0.08,0.5,NULL,NULL,NULL,0.04,NULL,0.04,NULL,3.1),
+(99,2,'Biscoito, polvilho doce',5.4,438,1.3,12.2,9,80.5,1.2,0.5,30,6,0.08,23,1.8,98,54,0.04,0.1,NULL,NULL,NULL,NULL,0.03,0.07,NULL,NULL),
+(100,2,'Brócolis, cozido',92.6,25,2.1,0.5,NULL,4.4,3.4,0.4,51,15,0.12,33,0.5,2,119,0.08,0.2,NULL,NULL,NULL,0.04,0.03,NULL,NULL,42),
+(101,2,'Brócolis, cru',91.2,25,3.6,0.3,NULL,4,2.9,0.8,86,30,0.26,78,0.6,3,322,0.06,0.5,NULL,279,140,0.12,0.18,0.08,1.39,34.3),
+(102,2,'Cará, cozido',78.9,78,1.5,0.1,NULL,18.9,2.6,0.6,5,15,0.02,28,0.3,1,203,0.11,0.2,NULL,NULL,NULL,0.12,NULL,0.12,NULL,NULL),
+(103,2,'Cará, cru',73.7,96,2.3,0.1,NULL,23,7.3,0.9,4,11,0.01,35,0.2,NULL,212,0.06,0.2,NULL,NULL,NULL,0.11,NULL,0.02,NULL,8.8),
+(104,2,'Caruru, cru',87.6,34,3.2,0.6,NULL,6,4.5,2.6,455,197,0.89,77,4.5,14,279,0.37,6,NULL,1906,953,NULL,0.1,0.11,NULL,5.4),
+(105,2,'Catalonha, crua',91.8,24,1.9,0.3,NULL,4.8,2,1.3,57,17,0.34,32,3.1,9,412,0.27,0.5,NULL,NULL,NULL,NULL,0.03,0.36,1.54,7.3),
+(106,2,'Catalonha, refogada',87.4,63,2,4.8,NULL,4.8,3.7,1,63,16,0.64,35,1.2,25,452,0.18,0.4,NULL,850,425,0.04,NULL,0.03,0.6,NULL),
+(107,2,'Cebola, crua',88.9,39,1.7,0.1,NULL,8.9,2.2,0.4,14,12,0.13,38,0.2,1,176,0.05,0.2,NULL,NULL,NULL,0.04,NULL,0.14,NULL,4.7),
+(108,2,'Cebolinha, crua',93.9,20,1.9,0.4,NULL,3.4,3.6,0.5,80,25,0.13,27,0.6,2,206,0.04,0.3,NULL,134,67,0.03,0.04,0.08,NULL,31.8),
+(109,2,'Cenoura, cozida',91.7,30,0.8,0.2,NULL,6.7,2.6,0.6,26,14,0.05,27,0.1,8,176,0.02,0.2,NULL,612,306,0.07,NULL,0.06,2.68,NULL),
+(110,2,'Cenoura, crua',90.1,34,1.3,0.2,NULL,7.7,3.2,0.9,23,11,0.05,28,0.2,3,315,0.05,0.2,NULL,1326,663,NULL,NULL,0.05,NULL,5.1),
+(111,2,'Chicória, crua',95.1,14,1.1,0.1,NULL,2.9,2.2,0.8,45,14,0.13,13,0.5,14,425,0.04,0.1,NULL,NULL,NULL,0.03,0.1,NULL,0.68,6.5),
+(112,2,'Chuchu, cozido',94.6,19,0.4,NULL,NULL,4.8,1,0.2,8,7,0.07,13,0.1,2,54,NULL,0.1,NULL,NULL,NULL,0.03,NULL,NULL,NULL,5.6),
+(113,2,'Chuchu, cru',94.8,17,0.7,0.1,NULL,4.1,1.3,0.3,12,7,0.08,18,0.2,NULL,126,0.03,0.1,NULL,NULL,NULL,NULL,NULL,NULL,NULL,10.6),
+(114,2,'Coentro, folhas desidratadas',10.6,309,20.9,10.4,NULL,48,37.3,10.2,784,393,10.48,388,81.4,18,3223,4.09,4.7,NULL,NULL,NULL,0.1,0.11,0.09,NULL,40.8),
+(115,2,'Couve, manteiga, crua',90.9,27,2.9,0.5,NULL,4.3,3.1,1.3,131,35,1.02,49,0.5,6,403,0.06,0.4,NULL,496,248,0.2,0.31,0.06,2.29,96.7),
+(116,2,'Couve, manteiga, refogada ',81.5,90,1.7,6.6,NULL,8.7,5.7,1.5,177,26,0.12,33,0.5,11,315,0.02,0.2,NULL,384,192,NULL,0.05,0.07,NULL,76.9),
+(117,2,'Couve-flor, crua',92.8,23,1.9,0.2,NULL,4.5,2.4,0.6,18,12,0.16,57,0.5,3,256,0.03,0.3,NULL,2,1,0.03,0.09,0.1,NULL,36.1),
+(118,2,'Couve-flor, cozida',94.3,19,1.2,0.3,NULL,3.9,2.1,0.3,16,5,0.1,25,0.1,2,80,NULL,0.3,NULL,NULL,NULL,0.04,NULL,NULL,NULL,23.7),
+(119,2,'Espinafre, Nova Zelândia, cru',94,16,2,0.2,NULL,2.6,2.1,1.2,98,82,0.71,25,0.4,17,336,0.06,0.3,NULL,484,242,0.1,0.21,0.06,NULL,2.4),
+(120,2,'Espinafre, Nova Zelândia, refogado',86.6,67,2.7,5.4,NULL,4.2,2.5,1,112,123,0.61,34,0.6,47,149,0.04,0.6,NULL,624,312,0.08,0.13,0.13,NULL,5.3),
+(121,2,'Farinha, de mandioca, crua',9.4,361,1.6,0.3,NULL,87.9,6.4,0.9,65,37,NULL,42,1.1,1,340,0.08,0.4,NULL,NULL,NULL,0.14,NULL,0.04,NULL,NULL),
+(122,2,'Farinha, de mandioca, torrada',8.3,365,1.2,0.3,NULL,89.2,6.5,1,76,40,0.37,39,1.2,10,328,NULL,0.4,NULL,NULL,NULL,NULL,NULL,0.81,NULL,NULL),
+(123,2,'Farinha, de puba',9.8,360,1.6,0.5,NULL,87.3,4.2,0.8,41,27,0.16,33,1.4,4,338,0.07,0.3,NULL,NULL,NULL,0.09,NULL,NULL,NULL,NULL),
+(124,2,'Fécula, de mandioca',17.8,331,0.5,0.3,NULL,81.1,0.6,0.3,12,3,NULL,60,0.1,2,48,NULL,NULL,NULL,NULL,NULL,0.03,NULL,NULL,NULL,NULL),
+(125,2,'Feijão, broto, cru',87.5,39,4.2,0.1,NULL,7.8,2,0.5,14,25,0.19,75,0.8,2,189,0.17,0.6,NULL,NULL,NULL,NULL,0.04,0.15,NULL,12),
+(126,2,'Inhame, cru',73.3,97,2.1,0.2,NULL,23.2,1.7,1.2,12,29,0.15,65,0.4,NULL,568,0.17,0.3,NULL,NULL,NULL,0.08,NULL,0.11,NULL,5.6),
+(127,2,'Jiló, cru',91.6,27,1.4,0.2,NULL,6.2,4.8,0.6,20,21,0.14,29,0.3,NULL,213,0.07,0.1,NULL,13,6,0.07,0.04,NULL,NULL,6.8),
+(128,2,'Jurubeba, crua',66.6,126,4.4,3.9,NULL,23.1,23.9,2,151,65,0.52,155,0.9,1,619,1.16,0.6,NULL,NULL,NULL,0.13,0.02,0.2,NULL,13.8),
+(129,2,'Mandioca, cozida',68.7,125,0.6,0.3,NULL,30.1,1.6,0.4,19,27,0.06,22,0.1,1,100,0.01,0.2,NULL,NULL,NULL,0.06,NULL,0.03,NULL,11.1),
+(130,2,'Mandioca, crua',61.8,151,1.1,0.3,NULL,36.2,1.9,0.6,15,44,0.05,29,0.3,2,208,0.07,0.2,NULL,NULL,NULL,NULL,NULL,0.04,NULL,16.5),
+(131,2,'Mandioca, farofa, temperada',6.4,406,2.1,9.1,NULL,80.3,7.8,2.1,66,34,0.29,45,1.4,575,201,NULL,0.2,NULL,NULL,NULL,0.1,NULL,0.15,NULL,NULL),
+(132,2,'Mandioca, frita',36.6,300,1.4,11.2,NULL,50.3,1.9,0.6,23,95,0.18,57,0.3,9,176,0.12,0.4,NULL,NULL,NULL,0.05,NULL,0.04,NULL,NULL),
+(133,2,'Manjericão, cru',93,21,2,0.4,NULL,3.6,3.3,1,211,58,0.17,40,1,4,252,0.16,0.5,NULL,1035,517,0.06,0.21,0.06,0.9,2.3),
+(134,2,'Maxixe, cru',95.1,14,1.4,0.1,NULL,2.7,2.2,0.7,21,10,0.07,25,0.4,11,328,0.02,0.2,NULL,6,3,0.06,0.02,0.04,NULL,9.6),
+(135,2,'Mostarda, folha, crua',93.4,18,2.1,0.2,NULL,3.2,1.9,1.1,68,16,0.14,58,1.1,3,364,0.05,0.3,NULL,NULL,NULL,0.05,0.04,0.05,NULL,38.6),
+(136,2,'Nhoque, batata, cozido',55,181,5.9,1.9,15,36.8,1.8,0.5,11,18,0.3,68,1.6,7,164,0.1,0.5,15,NULL,NULL,0.08,NULL,0.04,NULL,NULL),
+(137,2,'Nabo, cru',93.8,18,1.2,0.1,NULL,4.1,2.6,0.8,42,15,4.42,17,0.2,2,280,0.02,0.2,NULL,NULL,NULL,0.07,NULL,0.03,NULL,9.6),
+(138,2,'Palmito, juçara, em conserva',91.4,23,1.8,0.4,NULL,4.3,3.2,2.1,58,34,10.82,40,0.3,514,244,0.23,0.7,NULL,NULL,NULL,0.06,0.04,NULL,NULL,2),
+(139,2,'Palmito, pupunha, em conserva',89.4,29,2.5,0.5,NULL,5.5,2.6,2.1,32,25,0.14,55,0.2,563,206,0.08,0.4,NULL,NULL,NULL,0.03,NULL,NULL,NULL,8.7),
+(140,2,'Pão, de queijo, assado',33.7,363,5.1,24.6,68,34.2,0.6,2.3,102,8,0.03,94,0.3,773,93,0.01,0.6,61,NULL,NULL,0.04,0.1,0.04,NULL,NULL),
+(141,2,'Pão, de queijo, cru',41.8,295,3.6,14,63,38.5,1,2,88,7,NULL,79,0.3,405,58,NULL,0.4,48,NULL,NULL,NULL,0.08,NULL,NULL,NULL),
+(142,2,'Pepino, cru',96.8,10,0.9,NULL,NULL,2,1.1,0.3,10,9,0.08,12,0.1,NULL,154,0.04,0.1,NULL,4,2,NULL,0.03,NULL,NULL,5),
+(143,2,'Pimentão, amarelo, cru',91.9,28,1.2,0.4,NULL,6,1.9,0.5,10,11,0.08,22,0.4,NULL,221,0.04,0.2,NULL,40,20,0.04,0.03,0.06,NULL,201.4),
+(144,2,'Pimentão, verde, cru',93.5,21,1.1,0.2,NULL,4.9,2.6,0.4,9,8,0.14,17,0.4,NULL,174,0.07,0.1,NULL,46,23,NULL,NULL,NULL,NULL,100.2),
+(145,2,'Pimentão, vermelho, cru',92.9,23,1,0.1,NULL,5.5,1.6,0.4,6,11,0.06,20,0.3,NULL,211,0.04,0.2,NULL,86,43,0.05,0.06,0.02,NULL,158.2),
+(146,2,'Polvilho, doce',12.6,351,0.4,NULL,NULL,86.8,0.2,0.2,27,4,0.09,8,0.5,2,38,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(147,2,'Quiabo, cru',90.6,30,1.9,0.3,NULL,6.4,4.6,0.8,112,50,0.46,56,0.4,1,249,0.17,0.6,NULL,49,25,0.1,NULL,0.03,NULL,5.6),
+(148,2,'Rabanete, cru',95.1,14,1.4,0.1,NULL,2.7,2.2,0.7,21,10,0.07,25,0.4,11,328,0.02,0.2,NULL,NULL,NULL,0.06,0.02,0.04,NULL,9.6),
+(149,2,'Repolho, branco, cru',94.7,17,0.9,0.1,NULL,3.9,1.9,0.4,35,9,0.13,14,0.2,4,150,0.02,0.2,NULL,6,3,NULL,0.03,0.06,NULL,18.7),
+(150,2,'Repolho, roxo, cru',90.1,31,1.9,0.1,NULL,7.2,2,0.7,44,18,0.25,58,0.5,2,328,0.9,0.3,NULL,NULL,NULL,0.07,NULL,0.09,NULL,43.2),
+(151,2,'Repolho, roxo, refogado',88.7,42,1.8,1.2,NULL,7.6,1.8,0.7,43,17,0.26,59,0.5,3,321,0.02,0.3,NULL,NULL,NULL,0.07,NULL,0.07,0.63,40.5),
+(152,2,'Rúcula, crua',94.8,13,1.8,0.1,NULL,2.2,1.7,1.1,117,18,0.24,25,0.9,9,233,0.04,0.2,NULL,533,266,0.04,NULL,NULL,0.35,46.3),
+(153,2,'Salsa, crua',88.7,33,3.3,0.6,NULL,5.7,1.9,1.8,179,21,1.88,49,3.2,2,711,0.2,1.3,NULL,1743,872,0.12,0.15,0.47,0.72,51.7),
+(154,2,'Seleta de legumes, enlatada',82.1,57,3.4,0.4,NULL,12.7,3.1,1.4,16,16,0.13,49,1.1,398,122,0.08,0.5,NULL,244,122,0.03,NULL,NULL,3.72,NULL),
+(155,2,'Serralha, crua',90.2,30,2.7,0.7,NULL,4.9,3.5,1.4,126,30,0.23,48,1.3,19,265,0.2,1.3,NULL,1134,567,NULL,0.11,0.08,NULL,1.5),
+(156,2,'Taioba, crua',89.2,34,2.9,0.9,NULL,5.4,4.5,1.5,141,38,0.66,53,1.9,1,290,0.16,0.6,NULL,1160,580,NULL,0.1,0.1,NULL,17.9),
+(157,2,'Tomate, com semente, cru',95.1,15,1.1,0.2,NULL,3.1,1.2,0.5,7,11,0.07,20,0.2,1,222,0.04,0.1,NULL,54,27,0.12,NULL,0.02,NULL,21.2),
+(158,2,'Tomate, extrato',79.7,61,2.4,0.2,NULL,15,2.8,2.8,29,29,0.18,47,2.1,498,680,0.2,0.4,NULL,166,83,NULL,NULL,0.11,2.42,18),
+(159,2,'Tomate, molho industrializado',88.1,38,1.4,0.9,NULL,7.7,3.1,1.9,12,17,0.08,27,1.6,418,388,0.08,0.1,NULL,76,38,NULL,NULL,0.06,NULL,2.7),
+(160,2,'Tomate, purê',90.8,28,1.4,NULL,NULL,6.9,1,1,13,15,NULL,30,1.3,104,308,0.09,0.3,NULL,90,45,NULL,NULL,0.07,1.18,5.4),
+(161,2,'Tomate, salada',93.6,21,0.8,NULL,NULL,5.1,2.3,0.4,7,10,0.04,23,0.3,5,161,0.07,0.2,NULL,NULL,NULL,0.06,0.04,0.05,NULL,12.8),
+(162,2,'Vagem, crua',92.2,25,1.8,0.2,NULL,5.3,2.4,0.5,41,18,0.5,28,0.4,NULL,208,0.06,0.3,NULL,32,16,NULL,0.08,NULL,NULL,1.2),
+(163,4,'Abacate, cru',83.8,96,1.2,8.4,NULL,6,6.3,0.5,8,15,0.17,22,0.2,NULL,206,0.15,0.2,NULL,NULL,NULL,NULL,0.04,NULL,NULL,8.7),
+(164,4,'Abacaxi, cru',86.3,48,0.9,0.1,NULL,12.3,1,0.4,22,18,1.62,13,0.3,NULL,131,0.11,0.1,NULL,NULL,NULL,0.17,0.02,NULL,NULL,34.6),
+(165,4,'Abacaxi, polpa, congelada',91.3,31,0.5,0.1,NULL,7.8,0.3,0.3,14,10,1.02,8,0.4,1,107,0.04,0.1,NULL,2,1,0.05,NULL,NULL,NULL,1.2),
+(166,4,'Abiu, cru',83.1,62,0.8,0.7,NULL,14.9,1.7,0.4,6,9,0.08,20,0.2,NULL,128,0.09,0.1,NULL,NULL,NULL,NULL,0.04,0.05,NULL,10.3),
+(167,4,'Açaí, polpa, com xarope de guaraná e glucose',73.9,110,0.7,3.7,NULL,21.5,1.7,0.3,22,13,3.29,11,0.3,15,75,0.14,0.2,NULL,NULL,NULL,NULL,NULL,0.07,NULL,10.3),
+(168,4,'Açaí, polpa, congelada',88.7,58,0.8,3.9,NULL,6.2,2.6,0.3,35,17,6.16,16,0.4,5,124,0.18,0.3,NULL,NULL,NULL,NULL,0.04,0.07,NULL,NULL),
+(169,4,'Acerola, crua',90.5,33,0.9,0.2,NULL,8,1.5,0.4,13,13,0.07,9,0.2,NULL,165,0.07,0.1,NULL,247,124,NULL,0.04,NULL,1.38,941.4),
+(170,4,'Acerola, polpa, congelada',93.6,22,0.6,NULL,NULL,5.5,0.7,0.3,8,9,0.03,13,0.2,1,112,0.04,0.1,NULL,192,96,NULL,0.1,NULL,NULL,623.2),
+(171,4,'Ameixa, calda, enlatada ',52.2,183,0.4,NULL,NULL,46.9,0.5,0.5,13,10,0.1,21,2.2,3,221,0.03,0.1,NULL,NULL,NULL,NULL,0.23,0.37,NULL,4.3),
+(172,4,'Ameixa, crua',84.8,53,0.8,NULL,NULL,13.9,2.4,0.6,6,5,0.03,14,0.1,NULL,134,0.06,0.1,NULL,27,14,0.07,NULL,0.03,NULL,7.6),
+(173,4,'Ameixa, em calda, enlatada, drenada ',50.3,177,1,0.3,NULL,47.7,4.5,0.7,39,14,0.2,26,2.7,3,263,0.16,0.2,NULL,NULL,NULL,NULL,0.22,0.08,NULL,5.2),
+(174,4,'Atemóia, crua',72.7,97,1,0.3,NULL,25.3,2.1,0.7,23,25,0.16,23,0.2,1,300,0.17,0.2,NULL,NULL,NULL,0.09,0.07,0.07,1.58,10.1),
+(175,4,'Banana, da terra, crua',63.9,128,1.4,0.2,NULL,33.7,1.5,0.8,4,24,0.16,26,0.3,NULL,328,0.05,0.2,NULL,239,119,0.03,0.02,0.14,NULL,15.7),
+(176,4,'Banana, doce em barra',21.1,280,2.2,0.1,NULL,75.7,3.8,1,12,49,0.84,37,0.6,10,518,0.16,0.3,NULL,27,14,0.03,NULL,0.06,1.41,NULL),
+(177,4,'Banana, figo, crua',70.1,105,1.1,0.1,NULL,27.8,2.8,0.8,6,30,0.21,16,0.2,NULL,387,0.06,0.1,NULL,NULL,NULL,0.09,NULL,0.03,NULL,17.5),
+(178,4,'Banana, maçã, crua',75.2,87,1.8,0.1,NULL,22.3,2.6,0.6,3,24,0.6,29,0.2,NULL,264,0.11,0.1,NULL,6,3,NULL,NULL,0.14,NULL,10.5),
+(179,4,'Banana, nanica, crua',73.8,92,1.4,0.1,NULL,23.8,1.9,0.8,3,28,0.14,27,0.3,NULL,376,0.1,0.2,NULL,14,7,NULL,0.02,0.14,NULL,5.9),
+(180,4,'Banana, ouro, crua',68.2,112,1.5,0.2,NULL,29.3,2,0.8,3,28,0.09,22,0.3,NULL,355,0.08,0.3,NULL,50,25,NULL,NULL,0.14,NULL,7.6),
+(181,4,'Banana, pacova, crua',77.7,78,1.2,0.1,NULL,20.3,2,0.7,5,30,0.41,20,0.4,1,267,0.06,0.1,NULL,NULL,NULL,0.05,0.03,0.17,NULL,NULL),
+(182,4,'Banana, prata, crua',71.9,98,1.3,0.1,NULL,26,2,0.8,8,26,0.42,22,0.4,NULL,358,0.05,0.1,NULL,32,16,NULL,0.02,0.1,NULL,21.6),
+(183,4,'Cacau, cru',79.2,74,1,0.1,NULL,19.4,2.2,0.3,12,25,0.04,9,0.3,1,72,0.15,0.6,NULL,NULL,NULL,0.25,NULL,0.04,NULL,13.6),
+(184,4,'Cajá-Manga, cru',86.9,46,1.3,NULL,NULL,11.4,2.6,0.4,13,11,0.05,24,0.2,1,119,0.02,0.2,NULL,NULL,NULL,0.11,NULL,0.05,NULL,26.7),
+(185,4,'Cajá, polpa, congelada',92.4,26,0.6,0.2,NULL,6.4,1.4,0.4,9,7,0.07,26,0.3,7,148,0.1,0.1,NULL,106,53,0.04,NULL,NULL,NULL,NULL),
+(186,4,'Caju, cru',88.1,43,1,0.3,NULL,10.3,1.7,0.3,1,10,0.12,16,0.2,3,124,0.07,0.1,NULL,15,7,NULL,NULL,NULL,NULL,219.3),
+(187,4,'Caju, polpa, congelada',89.8,37,0.5,0.2,NULL,9.4,0.8,0.2,1,7,0.05,10,0.1,4,88,0.07,0.1,NULL,22,11,NULL,0.02,NULL,2.25,119.7),
+(188,4,'Caju, suco concentrado, envasado',88.4,45,0.4,0.2,NULL,10.7,0.6,0.3,1,8,0.06,11,0.1,45,107,0.04,0.1,NULL,10,5,NULL,0.03,NULL,NULL,138.7),
+(189,4,'Caqui, chocolate, cru',79.7,71,0.4,0.1,NULL,19.3,6.5,0.5,18,9,0.39,18,0.1,2,164,NULL,0.2,NULL,NULL,NULL,NULL,NULL,0.03,NULL,29.6),
+(190,4,'Carambola, crua',87.1,46,0.9,0.2,NULL,11.5,2,0.4,5,7,0.13,11,0.2,4,133,0.08,0.2,NULL,NULL,NULL,0.12,NULL,NULL,NULL,60.9),
+(191,4,'Ciriguela, crua',78.7,76,1.4,0.4,NULL,18.9,3.9,0.7,27,18,0.06,49,0.4,2,248,0.12,0.5,NULL,NULL,NULL,0.14,NULL,NULL,NULL,27),
+(192,4,'Cupuaçu, cru',86.2,49,1.2,1,NULL,10.4,3.1,1.2,13,18,0.07,21,0.5,3,331,0.07,0.3,NULL,NULL,NULL,0.37,0.04,0.07,4.34,24.5),
+(193,4,'Cupuaçu, polpa, congelada',86.6,49,0.8,0.6,NULL,11.4,1.6,0.6,5,14,0.17,14,0.3,1,291,0.14,0.2,NULL,NULL,NULL,0.07,0.07,0.05,NULL,10.5),
+(194,4,'Figo, cru',88.2,41,1,0.2,NULL,10.2,1.8,0.4,27,11,0.06,15,0.2,NULL,174,0.13,0.1,NULL,NULL,NULL,0.05,NULL,NULL,NULL,0.8),
+(195,4,'Figo, enlatado, em calda',48.8,184,0.6,0.2,NULL,50.3,2,0.2,33,7,0.16,6,0.5,7,39,0.25,0.1,NULL,NULL,NULL,NULL,NULL,NULL,NULL,5.2),
+(196,4,'Fruta-pão, crua',80.9,67,1.1,0.2,NULL,17.2,5.5,0.7,34,24,0.04,27,0.2,1,188,0.07,0.1,NULL,NULL,NULL,0.03,NULL,NULL,NULL,9.9),
+(197,4,'Goiaba, branca, com casca, crua',85.7,52,0.9,0.5,NULL,12.4,6.3,0.5,5,7,0.07,16,0.2,NULL,220,0.04,0.2,NULL,NULL,NULL,NULL,NULL,0.03,NULL,99.2),
+(198,4,'Goiaba, doce em pasta',24.8,269,0.6,0,NULL,74.1,3.7,0.5,10,6,0.11,54,0.4,4,165,0.06,0.1,NULL,136,68,0.07,NULL,NULL,NULL,23.1),
+(199,4,'Goiaba, doce, cascão',20.3,286,0.4,0.1,NULL,78.7,4.4,0.5,15,10,0.16,28,0.4,11,251,0.08,0.1,NULL,50,25,NULL,NULL,NULL,1.41,34.3),
+(200,4,'Goiaba, vermelha, com casca, crua',85,54,1.1,0.4,NULL,13,6.2,0.5,4,7,0.09,15,0.2,NULL,198,0.04,0.1,NULL,38,19,NULL,NULL,0.03,NULL,80.6),
+(201,4,'Graviola, crua',82.2,62,0.8,0.2,NULL,15.8,1.9,1,40,23,0.08,19,0.2,4,250,0.04,0.1,NULL,NULL,NULL,0.17,0.12,0.03,NULL,19.1),
+(202,4,'Graviola, polpa, congelada',89.2,38,0.6,0.1,NULL,9.8,1.2,0.4,6,10,0.06,17,0.1,3,170,0.06,0.1,NULL,NULL,NULL,NULL,0.09,NULL,NULL,10.5),
+(203,4,'Jabuticaba, crua',83.6,58,0.6,0.1,NULL,15.3,2.3,0.4,8,18,0.3,15,0.1,NULL,130,0.07,0.3,NULL,NULL,NULL,0.06,NULL,NULL,NULL,16.2),
+(204,4,'Jaca, crua',75.1,88,1.4,0.3,NULL,22.5,2.4,0.8,11,40,0.48,14,0.4,2,234,0.09,0.2,NULL,NULL,NULL,0.1,0.04,0.05,NULL,14.8),
+(205,4,'Jambo, cru',92.1,27,0.9,0.1,NULL,6.5,5.1,0.5,14,14,0.05,18,0.1,22,135,0.02,0.1,NULL,NULL,NULL,0.08,NULL,NULL,1.18,3.8),
+(206,4,'Jamelão, cru',87.7,41,0.5,0.1,NULL,10.6,1.8,1,3,2,NULL,4,0,1,394,0.03,0,NULL,8,4,0.17,NULL,0.12,NULL,27.1),
+(207,4,'Kiwi, cru',85.9,51,1.3,0.6,NULL,11.5,2.7,0.7,24,11,0.05,33,0.3,NULL,269,0.15,0.1,NULL,5,3,NULL,NULL,0.06,NULL,70.8),
+(208,4,'Laranja, baía, crua',87.1,45,1,0.1,NULL,11.5,1.1,0.4,35,9,0.04,24,0.1,NULL,174,0.04,0.1,NULL,4,2,0.06,0.02,NULL,NULL,56.9),
+(209,4,'Laranja, baía, suco',90.2,37,0.7,NULL,NULL,8.7,NULL,0.4,6,8,0.02,22,0.1,NULL,173,0.02,NULL,NULL,4,2,0.03,NULL,0.02,NULL,94.5),
+(210,4,'Laranja, da terra, crua',85.4,51,1.1,0.2,NULL,12.9,4,0.5,51,14,0.04,20,0.1,1,173,0.04,0.2,NULL,NULL,NULL,0.07,0.09,0.03,NULL,34.7),
+(211,4,'Laranja, da terra, suco',89.2,41,0.7,0.1,NULL,9.6,1,0.4,13,10,0.02,15,0.1,NULL,145,0.02,0.1,NULL,4,2,0.04,NULL,0.04,NULL,44.3),
+(212,4,'Laranja, lima, crua',87,46,1.1,0.1,NULL,11.5,1.8,0.4,31,10,0.05,15,0.1,1,130,0.03,0.1,NULL,NULL,NULL,0.09,0.05,0.04,NULL,43.5),
+(213,4,'Laranja, lima, suco',89.7,39,0.7,0.1,NULL,9.2,0.4,0.3,8,11,0.02,16,NULL,NULL,129,0.02,0,NULL,NULL,NULL,0.1,0.03,0.04,NULL,41.3),
+(214,4,'Laranja, pêra, crua',89.6,37,1,0.1,NULL,8.9,0.8,0.3,22,9,0.05,23,0.1,NULL,163,0.03,0.1,NULL,1,0,0.07,0.02,0.02,NULL,53.7),
+(215,4,'Laranja, pêra, suco',91.3,33,0.7,0.1,NULL,7.6,NULL,0.3,7,8,0.03,14,NULL,NULL,149,0.01,NULL,NULL,1,1,NULL,NULL,NULL,NULL,73.3),
+(216,4,'Laranja, valência, crua',86.9,46,0.8,0.2,NULL,11.7,1.7,0.4,34,14,0.06,20,0.1,1,158,0.04,0.1,NULL,NULL,NULL,0.07,0.04,0.03,NULL,47.8),
+(217,4,'Laranja, valência, suco',90.5,36,0.5,0.1,NULL,8.6,0.4,0.3,9,10,0.03,17,NULL,NULL,143,0.02,NULL,NULL,2,1,NULL,NULL,0.03,NULL,NULL),
+(218,4,'Limão, cravo, suco',94.2,14,0.3,NULL,NULL,5.2,NULL,0.2,10,9,0.03,11,0.1,NULL,120,0.03,NULL,NULL,31,16,NULL,NULL,NULL,NULL,32.8),
+(219,4,'Limão, galego, suco',91.8,22,0.6,0.1,NULL,7.3,NULL,0.3,5,6,0.01,13,0.1,NULL,113,0.02,0.1,NULL,NULL,NULL,NULL,NULL,0.03,NULL,34.5),
+(220,4,'Limão, tahiti, cru',87.4,32,0.9,0.1,NULL,11.1,1.2,0.4,51,10,0.07,24,0.2,1,128,0.06,0.2,NULL,NULL,NULL,0.3,0.04,NULL,NULL,38.2),
+(221,4,'Maçã, Argentina, com casca, crua',82.6,63,0.2,0.2,NULL,16.6,2,0.3,3,5,0.01,11,0.1,1,117,0.03,NULL,NULL,NULL,NULL,0.05,NULL,NULL,NULL,1.5),
+(222,4,'Maçã, Fuji, com casca, crua',84.3,56,0.3,NULL,NULL,15.2,1.3,0.2,2,2,0.03,9,0.1,NULL,75,0.06,NULL,NULL,4,2,NULL,NULL,0.03,NULL,2.4),
+(223,4,'Macaúba, crua',41.5,404,2.1,40.7,NULL,13.9,13.4,1.8,67,66,0.08,44,0.8,1,306,0.35,0.7,NULL,1020,510,0.11,0.11,0.03,NULL,13.4),
+(224,4,'Mamão, doce em calda, drenado',45.5,196,0.2,0.1,NULL,54,1.3,0.2,20,6,0.02,4,0.1,3,68,0.02,0.1,NULL,NULL,NULL,0.03,NULL,NULL,NULL,3.9),
+(225,4,'Mamão, Formosa, cru',86.9,45,0.8,0.1,NULL,11.6,1.8,0.6,25,17,0.04,11,0.2,3,222,1.36,0.1,NULL,148,74,0.03,0.03,NULL,NULL,78.5),
+(226,4,'Mamão, Papaia, cru',88.6,40,0.5,0.1,NULL,10.4,1,0.4,22,22,0.01,11,0.2,2,126,0.02,0.1,NULL,118,59,0.03,0.04,NULL,1.03,82.2),
+(227,4,'Mamão verde, doce em calda, drenado',41.9,209,0.3,0.1,NULL,57.6,1.2,0.1,12,5,0.04,3,0.2,5,9,0.02,0,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(228,4,'Manga, Haden, crua',82.3,64,0.4,0.3,NULL,16.7,1.6,0.4,12,8,0.17,9,0.1,1,148,0.1,0.1,NULL,162,81,0.02,0.06,0.05,NULL,17.4),
+(229,4,'Manga, Palmer, crua',79.7,72,0.4,0.2,NULL,19.4,1.6,0.3,12,9,0.05,14,0.1,2,157,0.09,0.1,NULL,787,393,0.09,0.03,NULL,NULL,65.5),
+(230,4,'Manga, polpa, congelada',86.5,48,0.4,0.2,NULL,12.5,1.1,0.4,7,9,0.12,9,0.1,7,131,0.06,0.1,NULL,NULL,NULL,NULL,0.07,0.03,NULL,24.9),
+(231,4,'Manga, Tommy Atkins, crua',85.8,51,0.9,0.2,NULL,12.8,2.1,0.3,8,7,0.34,14,0.1,NULL,138,0.06,0.1,NULL,118,59,NULL,0.04,0.03,NULL,7.9),
+(232,4,'Maracujá, cru',82.9,68,2,2.1,NULL,12.3,1.1,0.8,5,28,0.12,51,0.6,2,338,0.19,0.4,NULL,114,57,NULL,0.05,0.05,NULL,19.8),
+(233,4,'Maracujá, polpa, congelada',88.9,39,0.8,0.2,NULL,9.6,0.5,0.5,5,10,0.07,15,0.3,8,228,0.05,0.2,NULL,NULL,NULL,NULL,0.09,0.06,NULL,7.3),
+(234,4,'Maracujá, suco concentrado, envasado',88.9,42,0.8,0.2,NULL,9.6,0.4,0.5,4,4,0.07,14,0.3,22,201,0.05,0.1,NULL,NULL,NULL,NULL,0.08,0.05,1.92,13.7),
+(235,4,'Melancia, crua',90.7,33,0.9,NULL,NULL,8.1,0.1,0.3,8,10,0.14,12,0.2,NULL,104,0.04,0.1,NULL,61,31,NULL,NULL,NULL,NULL,6.1),
+(236,4,'Melão, cru',91.3,29,0.7,NULL,NULL,7.5,0.3,0.5,3,6,0.05,10,0.2,11,216,0.04,0.1,NULL,2,1,NULL,NULL,0.02,NULL,8.7),
+(237,4,'Mexerica, Murcote, crua',83.7,58,0.9,0.1,NULL,14.9,3.1,0.5,33,12,0.05,19,0.1,1,159,0.06,0.1,NULL,NULL,NULL,0.08,0.07,0.05,NULL,21.8),
+(238,4,'Mexerica, Rio, crua',89.6,37,0.7,0.1,NULL,9.3,2.7,0.3,17,8,0.05,14,0.1,2,125,0.07,0.2,NULL,NULL,NULL,0.03,NULL,NULL,NULL,112),
+(239,4,'Morango, cru',91.5,30,0.9,0.3,NULL,6.8,1.7,0.5,11,10,0.33,22,0.3,NULL,184,0.06,0.2,NULL,NULL,NULL,NULL,0.03,0.03,NULL,63.6),
+(240,4,'Nêspera, crua',87.8,43,0.3,NULL,NULL,11.5,3,0.4,20,10,0.07,10,0.1,NULL,113,0.12,0.1,NULL,170,85,NULL,NULL,NULL,NULL,3.2),
+(241,4,'Pequi, cru',65.9,205,2.3,18,NULL,13,19,0.8,32,30,0.64,34,0.3,NULL,298,0.21,1,NULL,60,30,0.17,0.48,0.06,2.57,8.3),
+(242,4,'Pêra, Park, crua',83.2,61,0.2,0.2,NULL,16.1,3,0.3,9,6,0.04,12,0.3,1,102,0.08,0.1,NULL,NULL,NULL,0.06,NULL,NULL,NULL,2.4),
+(243,4,'Pêra, Williams, crua',85,53,0.6,0.1,NULL,14,3,0.3,8,6,0.04,12,0.1,NULL,116,0.07,0.1,NULL,NULL,NULL,NULL,NULL,NULL,NULL,2.8),
+(244,4,'Pêssego, Aurora, cru',89.3,36,0.8,NULL,NULL,9.3,1.4,0.5,3,4,0.05,15,0.2,NULL,124,0.02,0.1,NULL,NULL,NULL,0.05,NULL,NULL,NULL,3.3),
+(245,4,'Pêssego, enlatado, em calda',82.2,63,0.7,NULL,NULL,16.9,1,0.2,4,4,0.03,9,0.6,3,95,0.04,0.1,NULL,48,24,NULL,NULL,NULL,NULL,NULL),
+(246,4,'Pinha, crua',75,88,1.5,0.3,NULL,22.4,3.4,0.7,21,31,0.15,34,0.2,1,283,0.11,0.2,NULL,NULL,NULL,0.12,0.04,0.09,NULL,35.9),
+(247,4,'Pitanga, crua',88.3,41,0.9,0.2,NULL,10.2,3.2,0.4,18,12,0.36,13,0.4,2,113,0.08,0.4,NULL,154,77,0.03,0.1,NULL,NULL,24.9),
+(248,4,'Pitanga, polpa, congelada',94.6,19,0.3,0.1,NULL,4.8,0.7,0.3,8,6,0.05,12,0.4,5,87,0.06,0.1,NULL,146,73,NULL,NULL,NULL,NULL,NULL),
+(249,4,'Romã, crua',84,56,0.4,NULL,NULL,15.1,0.4,0.5,5,13,0.13,40,0.3,1,485,0.19,0.7,NULL,NULL,NULL,0.12,0.17,0.05,NULL,8.1),
+(250,4,'Tamarindo, cru',22,276,3.2,0.5,NULL,72.5,6.4,1.9,37,59,0.34,55,0.6,0,723,0.29,0.7,NULL,NULL,NULL,0.31,NULL,0.1,NULL,7.2),
+(251,4,'Tangerina, Poncã, crua',89.2,38,0.8,0.1,NULL,9.6,0.9,0.3,13,8,0.04,12,0.1,NULL,131,0.03,NULL,NULL,47,24,0.06,0.02,0.02,NULL,48.8),
+(252,4,'Tangerina, Poncã, suco',90.4,36,0.5,NULL,NULL,8.8,NULL,0.3,4,6,0.02,9,NULL,NULL,119,0.02,NULL,NULL,51,26,0.06,NULL,0.03,NULL,41.8),
+(253,4,'Tucumã, cru',51.3,262,2.1,19.1,NULL,26.5,12.7,1.1,46,121,0.55,53,0.6,4,401,0.39,0.9,NULL,2363,1181,0.04,NULL,0.11,NULL,18),
+(254,4,'Umbu, cru',89.3,37,0.8,NULL,NULL,9.4,2,0.5,12,11,0.03,13,0.1,NULL,152,0.04,0.4,NULL,NULL,NULL,0.05,0.06,0.03,NULL,24.1),
+(255,4,'Umbu, polpa, congelada',90.2,34,0.5,0.1,NULL,8.8,1.3,0.4,11,8,0.05,13,0.2,6,154,0.04,0.1,NULL,NULL,NULL,NULL,0.06,NULL,NULL,4),
+(256,4,'Uva, Itália, crua',85,53,0.7,0.2,NULL,13.6,0.9,0.6,7,5,0.13,12,0.1,NULL,162,0.11,NULL,NULL,5,3,NULL,NULL,0.03,NULL,3.3),
+(257,4,'Uva, Rubi, crua',86.1,49,0.6,0.2,NULL,12.7,0.9,0.5,8,6,0.07,23,0.2,8,159,0.05,NULL,NULL,NULL,NULL,0.02,0.02,NULL,NULL,1.9),
+(258,4,'Uva, suco concentrado, envasado',85.1,58,NULL,NULL,NULL,14.7,0.2,0.2,9,7,0.2,10,0.1,10,54,0.13,0.1,NULL,NULL,NULL,0.07,0.02,0.05,NULL,21),
+(259,3,'Azeite, de dendê',NULL,884,NULL,100,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(260,3,'Azeite, de oliva, extra virgem',NULL,884,NULL,100,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(261,3,'Manteiga, com sal',15.8,726,0.4,82.4,201,0.1,NULL,1.4,9,1,NULL,28,0.2,579,15,NULL,NULL,924,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(262,3,'Manteiga, sem sal',13.6,758,0.4,86,214,0,NULL,0.1,4,1,NULL,7,NULL,4,5,0.04,NULL,1013,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(263,3,'Margarina, com óleo hidrogenado, com sal (65% de lipídeos)',32.2,596,NULL,67.4,NULL,0,NULL,1.9,6,1,NULL,7,0.1,894,21,NULL,NULL,462,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(264,3,'Margarina, com óleo hidrogenado, sem sal (80% de lipídeos)',19.6,723,NULL,81.7,NULL,0,NULL,0.2,3,1,NULL,4,0.1,78,2,NULL,NULL,534,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(265,3,'Margarina, com óleo interesterificado, com sal (65% de lipídeos)',32,594,NULL,67.2,NULL,0,NULL,1.2,5,1,NULL,6,NULL,561,15,NULL,NULL,385,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(266,3,'Margarina, com óleo interesterificado, sem sal (65% de lipídeos)',33.4,593,NULL,67.1,NULL,0,NULL,0.1,5,1,NULL,7,0.1,33,5,NULL,NULL,245,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(267,3,'Óleo, de babaçu',NULL,884,NULL,100,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(268,3,'Óleo, de canola',NULL,884,NULL,100,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(269,3,'Óleo, de girassol',NULL,884,NULL,100,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(270,3,'Óleo, de milho',NULL,884,NULL,100,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(271,3,'Óleo, de pequi',NULL,884,NULL,100,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(272,3,'Óleo, de soja',NULL,884,NULL,100,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(273,5,'Abadejo, filé, congelado, assado',74.3,112,23.5,1.2,103,0,NULL,1.6,23,20,NULL,338,0.5,334,156,0.03,0.5,NULL,NULL,NULL,NULL,NULL,NULL,0.44,NULL),
+(274,5,'Abadejo, filé, congelado, cozido',79.7,91,19.3,0.9,87,0,NULL,0.9,17,16,NULL,351,0.3,189,146,0.03,0.4,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(275,5,'Abadejo, filé, congelado, cru',86.4,59,13.1,0.4,31,0,NULL,0.6,10,14,0.01,91,0.1,79,148,0.14,0.4,NULL,NULL,NULL,0.04,NULL,NULL,4.6,NULL),
+(276,5,'Abadejo, filé, congelado, grelhado',71,130,27.6,1.3,136,0,NULL,1.6,20,22,NULL,581,0.3,305,279,NULL,0.4,6,6,6,0.03,NULL,NULL,0.44,NULL),
+(277,5,'Atum, conserva em óleo',64.5,166,26.2,6,53,0,NULL,1.5,7,29,NULL,211,1.2,362,280,0.04,0.6,NULL,NULL,NULL,0.15,0.03,NULL,3.17,NULL),
+(278,5,'Atum, fresco, cru',73.1,118,25.7,0.9,48,0,NULL,1.3,7,32,NULL,254,1.3,30,308,0.09,0.4,20,20,20,NULL,0.04,NULL,5.94,NULL),
+(279,5,'Bacalhau, salgado, cru',47.9,136,29,1.3,139,0,NULL,22.5,157,49,0.03,186,0.9,13585,434,0.09,0.7,NULL,NULL,NULL,NULL,0.21,NULL,NULL,NULL),
+(280,5,'Bacalhau, salgado, refogado',65.9,140,24,3.6,112,1.2,NULL,5.3,59,15,NULL,51,0.2,1256,50,0.04,0.6,NULL,NULL,NULL,0.03,0.05,NULL,5.17,NULL),
+(281,5,'Cação, posta, com farinha de trigo, frita',60.6,208,25,10,75,3.1,0.5,1.4,30,26,0.06,462,1,160,420,NULL,0.6,17,17,17,NULL,NULL,NULL,0.77,NULL),
+(282,5,'Cação, posta, cozida',75.9,116,25.6,0.7,83,0,NULL,1.2,10,21,NULL,204,0.3,115,249,0.03,0.6,12,12,12,NULL,0.04,NULL,9.77,NULL),
+(283,5,'Cação, posta, crua',81.4,83,17.9,0.8,36,0,NULL,1.2,9,19,0.04,181,0.2,176,299,0.13,0.3,6,6,6,0.04,NULL,NULL,1.31,NULL),
+(284,5,'Camarão, Rio Grande, grande, cozido',78.7,90,19,1,241,0,NULL,1.7,90,19,0.06,266,1.3,367,102,0.17,1.2,NULL,NULL,NULL,NULL,NULL,NULL,1.07,NULL),
+(285,5,'Camarão, Rio Grande, grande, cru',89.1,47,10,0.5,124,0,NULL,0.8,51,27,0.04,234,0.7,201,72,0.11,0.7,20,20,20,NULL,NULL,NULL,NULL,NULL),
+(286,5,'Camarão, Sete Barbas, sem cabeça, com casca, frito',61,231,18.4,15.6,283,2.9,NULL,2.1,960,74,0.38,337,2.4,99,107,0.19,1.1,NULL,NULL,NULL,0.05,NULL,NULL,0.35,NULL),
+(287,5,'Caranguejo, cozido',77,83,18.5,0.4,85,0,NULL,3.5,357,52,0.07,154,2.9,360,186,0.72,5.7,NULL,NULL,NULL,0.04,0.31,NULL,4.17,NULL),
+(288,5,'Corimba, cru',75.6,128,17.4,6,40,0,NULL,1,40,23,0.02,190,0.5,47,317,0.03,0.4,NULL,NULL,NULL,NULL,NULL,NULL,1.87,NULL),
+(289,5,'Corimbatá, assado',59.9,261,19.9,19.6,80,0,NULL,1,22,24,0.01,221,1,40,326,0.03,0.7,NULL,NULL,NULL,0.13,0.06,NULL,2.7,NULL),
+(290,5,'Corimbatá, cozido',64.6,239,20.1,16.9,75,0,NULL,0.8,65,23,0.02,185,0.6,37,254,0.05,1,NULL,NULL,NULL,NULL,NULL,NULL,2.09,NULL),
+(291,5,'Corvina de água doce, crua',79.2,101,18.9,2.2,73,0,NULL,1,39,25,0.02,154,0.3,45,293,0.02,0.4,8,8,8,NULL,0.08,NULL,NULL,NULL),
+(292,5,'Corvina do mar, crua',79.4,94,18.6,1.6,67,0,NULL,1.1,NULL,24,0.01,183,0.4,68,339,0.03,0.4,65,65,65,0.12,0.05,NULL,2.62,NULL),
+(293,5,'Corvina grande, assada',69,147,26.8,3.6,117,0,NULL,1.3,60,24,0.03,176,0.5,85,291,0.02,0.7,NULL,NULL,NULL,0.11,0.1,NULL,4.23,NULL),
+(294,5,'Corvina grande, cozida',73.6,100,23.4,2.6,123,0,NULL,0.9,69,22,0.02,164,0.6,68,194,0.03,0.7,6,6,6,NULL,0.03,NULL,0.77,NULL),
+(295,5,'Dourada de água doce, fresca',76.2,131,18.8,5.6,52,0,NULL,1.1,12,26,0.01,189,0.2,40,393,0.02,0.5,NULL,NULL,NULL,NULL,NULL,NULL,3.7,NULL),
+(296,5,'Lambari, congelado, cru',71.9,131,16.8,6.5,159,0,NULL,3.6,1181,45,0.28,696,0.9,48,244,0.09,3.3,4,4,4,0.05,NULL,NULL,3.3,NULL),
+(297,5,'Lambari, congelado, frito',40.1,327,28.4,22.8,246,0,NULL,6.2,1881,66,0.34,1067,0.8,65,331,0.31,5.6,9,9,9,0.25,0.03,0.04,8.93,NULL),
+(298,5,'Lambari, fresco, cru',72.2,152,15.7,9.4,144,0,NULL,2.2,590,32,0.41,441,0.6,41,207,0.05,2.4,NULL,NULL,NULL,NULL,0.03,0.07,1.86,NULL),
+(299,5,'Manjuba, com farinha de trigo, frita',41.3,344,23.5,22.6,282,10.2,0.4,2.4,763,47,0.36,578,3,37,319,0.15,3.8,30,30,30,0.05,0.07,NULL,0.79,NULL),
+(300,5,'Manjuba, frita',40.7,349,30.1,24.5,270,0,NULL,4.2,575,32,0.21,735,0.9,41,318,0.14,3.2,12,12,12,0.03,0.03,NULL,7.27,NULL),
+(301,5,'Merluza, filé, assado',70.7,122,26.6,0.9,91,0,NULL,1.2,36,20,0.03,273,0.4,120,364,0.03,0.9,NULL,NULL,NULL,0.05,NULL,NULL,7.97,NULL),
+(302,5,'Merluza, filé, cru',82.1,89,16.6,2,57,0,NULL,1.1,20,27,0.01,185,0.2,80,340,NULL,0.3,NULL,NULL,NULL,NULL,NULL,NULL,1.34,NULL),
+(303,5,'Merluza, filé, frito',63.5,192,26.9,8.5,109,0,NULL,1.5,36,38,0.02,279,0.4,90,447,0.03,0.6,NULL,NULL,NULL,0.05,NULL,NULL,0.77,NULL),
+(304,5,'Pescada, branca, crua',79.6,111,16.3,4.6,51,0,NULL,0.9,16,19,0.01,136,0.2,76,261,NULL,0.3,3,3,3,NULL,0.04,NULL,0.6,NULL),
+(305,5,'Pescada, branca, frita',57,223,27.4,11.8,165,0,NULL,3.2,378,30,0.06,504,0.5,107,355,0.08,1.1,39,39,39,0.08,0.11,NULL,8.07,NULL),
+(306,5,'Pescada, filé, com farinha de trigo, frito',53.2,283,21.4,19.1,73,5,NULL,1.2,26,28,0.07,216,0.9,91,216,0.03,0.4,18,18,18,NULL,0.03,NULL,0.6,NULL),
+(307,5,'Pescada, filé, cru',79.5,107,16.7,4,65,0,NULL,0.9,14,23,0.01,141,0.2,77,253,NULL,0.3,48,48,48,NULL,0.05,NULL,3.1,NULL),
+(308,5,'Pescada, filé, frito',66.8,154,28.6,3.6,81,0,NULL,1.7,10,21,NULL,204,0.3,115,249,0.03,0.6,4,4,4,0.14,NULL,NULL,8.87,NULL),
+(309,5,'Pescada, filé, molho escabeche',74.5,142,11.8,8,43,5,0.8,0.7,20,18,0.08,105,1.5,51,208,0.36,0.3,19,NULL,NULL,NULL,0.03,NULL,1.1,6.9),
+(310,5,'Pescadinha, crua',80.6,76,15.5,1.1,84,0,NULL,2,332,34,0.07,327,0.5,120,304,0.1,0.6,NULL,NULL,NULL,0.08,NULL,NULL,6.6,NULL),
+(311,5,'Pintado, assado',57,192,36.5,4,126,0,NULL,2,114,42,0.08,332,0.8,81,527,0.04,2.1,7,7,7,0.03,NULL,NULL,6.63,NULL),
+(312,5,'Pintado, cru',80.3,91,18.6,1.3,50,0,NULL,1.1,12,24,0.01,174,0.2,43,294,0.03,0.4,NULL,NULL,NULL,NULL,NULL,NULL,4.95,NULL),
+(313,5,'Pintado, grelhado',65.5,152,30.8,2.3,99,0,NULL,1.6,69,27,0.03,237,0.5,53,360,0.04,0.8,5,5,5,0.03,NULL,NULL,2.87,NULL),
+(314,5,'Porquinho, cru',79.2,93,20.5,0.6,49,0,NULL,1.3,26,24,0.05,207,0.4,67,313,0.04,0.7,5,5,5,NULL,NULL,NULL,1.19,NULL),
+(315,5,'Salmão, filé, com pele, fresco,  grelhado',60.9,229,23.9,14,85,0,NULL,1.3,29,28,0.02,300,0.5,85,384,0.04,0.6,NULL,NULL,NULL,0.13,0.06,NULL,5.35,NULL),
+(316,5,'Salmão, sem pele, fresco, cru',69,170,19.3,9.7,53,0,NULL,1.2,9,27,NULL,259,0.2,64,376,0.02,0.3,NULL,NULL,NULL,0.2,0.03,NULL,3.21,NULL),
+(317,5,'Salmão, sem pele, fresco, grelhado',58.1,243,26.1,14.5,73,0,NULL,1.6,15,38,0.01,352,0.4,96,518,0.04,0.5,NULL,NULL,NULL,0.23,0.03,NULL,6.3,NULL),
+(318,5,'Sardinha, assada',60.1,164,32.2,3,109,0,NULL,4.4,438,51,0.24,578,1.3,74,574,0.14,1.8,NULL,NULL,NULL,0.06,NULL,NULL,5.83,NULL),
+(319,5,'Sardinha, conserva em óleo',55.1,285,15.9,24,73,0,NULL,2.9,550,35,0.11,496,3.5,666,367,0.03,1.6,NULL,NULL,NULL,0.42,0.04,NULL,6.55,NULL),
+(320,5,'Sardinha, frita',48.5,257,33.4,12.7,103,0,NULL,4.3,482,39,0.25,629,1.1,60,460,0.14,1.6,NULL,NULL,NULL,0.06,NULL,NULL,7.1,NULL),
+(321,5,'Sardinha, inteira, crua',76.6,114,21.1,2.7,61,0,NULL,1.6,167,29,0.1,294,1.3,60,312,0.13,1.3,NULL,NULL,NULL,NULL,0.07,NULL,7.7,NULL),
+(322,5,'Tucunaré, filé, congelado, cru',79.9,88,18,1.2,47,0,NULL,1,19,26,0.01,168,0.3,57,288,0.1,0.4,NULL,NULL,NULL,NULL,0.05,NULL,0.9,NULL),
+(323,6,'Apresuntado',73.7,129,13.5,6.7,38,2.9,NULL,3.3,23,15,0.04,225,0.9,943,270,0.03,1.6,NULL,NULL,NULL,0.62,0.04,NULL,0.7,NULL),
+(324,6,'Caldo de carne, tablete',2.9,241,7.8,16.6,NULL,15.1,0.6,57.6,129,22,0.02,123,NULL,22180,218,NULL,NULL,NULL,NULL,NULL,0.7,0.07,0.27,NULL,NULL),
+(325,6,'Caldo de galinha, tablete',3.4,251,6.3,20.4,2,10.6,11.8,59.3,16,13,0.13,48,0.7,22300,68,NULL,0.3,NULL,NULL,NULL,0.05,0.04,NULL,NULL,NULL),
+(326,6,'Carne, bovina, acém, moído, cozido',61.6,212,26.7,10.9,103,0,NULL,0.8,4,17,0.01,164,2.7,52,256,0.06,8.1,NULL,NULL,NULL,NULL,0.32,NULL,1.76,NULL),
+(327,6,'Carne, bovina, acém, moído, cru',72.7,137,19.4,5.9,58,0,NULL,0.9,3,14,NULL,158,1.8,49,237,0.08,6.3,2,2,2,0.15,0.21,0.03,4.33,NULL),
+(328,6,'Carne, bovina, acém, sem gordura, cozido',60.4,215,27.3,10.9,107,0,NULL,0.8,7,14,NULL,164,2.4,56,254,0.07,8,NULL,NULL,NULL,NULL,0.04,0.07,1.63,NULL),
+(329,6,'Carne, bovina, acém, sem gordura, cru',71.5,144,20.8,6.1,53,0,NULL,1,5,13,NULL,144,1.5,50,234,0.04,5.2,2,2,2,0.12,0.12,NULL,2.33,NULL),
+(330,6,'Carne, bovina, almôndegas, cruas',64.3,189,12.3,11.2,34,9.8,NULL,2.4,22,24,0.17,145,1.6,621,328,0.15,2.3,NULL,NULL,NULL,0.11,0.04,NULL,2.2,NULL),
+(331,6,'Carne, bovina, almôndegas, fritas',48.1,272,18.2,15.8,36,14.3,NULL,3.7,27,48,0.41,244,1.9,1030,536,0.19,2.6,NULL,NULL,NULL,0.13,0.07,NULL,6.6,NULL),
+(332,6,'Carne, bovina, bucho, cozido',74.1,133,21.6,4.5,245,0,NULL,0.3,13,7,0.01,63,0.6,38,70,0.05,2.5,NULL,NULL,NULL,NULL,NULL,NULL,2.18,NULL),
+(333,6,'Carne, bovina, bucho, cru',75,137,20.5,5.5,145,0,NULL,0.4,9,6,0.01,61,0.5,45,85,0.06,2.1,NULL,NULL,NULL,NULL,NULL,NULL,2.13,NULL),
+(334,6,'Carne, bovina, capa de contra-filé, com gordura, crua',64.8,217,19.2,15,63,0,NULL,0.9,6,17,0.01,144,1.5,58,267,0.06,3.5,4,4,4,0.09,0.09,0.12,1.55,NULL),
+(335,6,'Carne, bovina, capa de contra-filé, com gordura, grelhada',47.9,312,30.7,20,120,0,NULL,1.1,7,18,NULL,214,2.6,81,323,0.13,6.2,NULL,NULL,NULL,NULL,0.06,NULL,1.74,NULL),
+(336,6,'Carne, bovina, capa de contra-filé, sem gordura, crua',73,131,21.5,4.3,58,0,NULL,1,6,20,NULL,178,2,79,325,0.06,4.6,NULL,NULL,NULL,NULL,0.04,NULL,1.81,NULL),
+(337,6,'Carne, bovina, capa de contra-filé, sem gordura, grelhada',53.7,239,35.1,10,80,0,NULL,1.3,9,26,0.01,287,2.8,83,385,0.12,7.6,NULL,NULL,NULL,0.03,0.08,NULL,1.86,NULL),
+(338,6,'Carne, bovina, charque, cozido',45.8,263,36.4,11.9,113,0,NULL,3.6,15,13,0.02,101,3.5,1443,90,0.07,6.1,NULL,NULL,NULL,0.05,0.07,NULL,1.5,NULL),
+(339,6,'Carne, bovina, charque, cru',44.5,249,22.7,16.8,81,0,NULL,14.5,15,13,NULL,122,1.5,5875,236,0.03,3.9,NULL,NULL,NULL,0.12,0.07,NULL,1.63,NULL),
+(340,6,'Carne, bovina, contra-filé, à milanesa',42.2,352,20.6,24,99,12.2,0.4,1.1,15,27,0.27,203,2.9,77,271,0.1,2.9,15,NULL,NULL,0.07,0.04,NULL,3.02,NULL),
+(341,6,'Carne, bovina, contra-filé de costela, cru',66.4,202,19.8,13.1,52,0,NULL,1,3,14,NULL,164,1.6,39,245,0.04,4.4,3,3,3,0.14,0.08,NULL,2.68,NULL),
+(342,6,'Carne, bovina, contra-filé de costela, grelhado',52.2,275,29.9,16.3,98,0,NULL,1.2,4,24,0.01,252,2.8,51,383,0.08,6.7,NULL,NULL,NULL,NULL,0.19,0.17,2.75,NULL),
+(343,6,'Carne, bovina, contra-filé, com gordura, cru',65.7,206,21.2,12.8,73,0,NULL,0.9,4,18,0.01,164,1.3,44,285,0.04,2.8,4,4,4,0.11,0.08,0.03,3.79,NULL),
+(344,6,'Carne, bovina, contra-filé, com gordura, grelhado',51.7,278,32.4,15.5,144,0,NULL,1.2,4,19,NULL,219,2.4,57,352,0.09,4.8,NULL,NULL,NULL,NULL,0.18,0.13,4.91,NULL),
+(345,6,'Carne, bovina, contra-filé, sem gordura, cru',69.1,157,24,6,59,0,NULL,1,4,21,NULL,184,1.7,53,335,0.05,3.2,NULL,NULL,NULL,NULL,0.2,NULL,4.64,NULL),
+(346,6,'Carne, bovina, contra-filé, sem gordura, grelhado',57.5,194,35.9,4.5,102,0,NULL,1.3,5,21,NULL,241,2.4,58,386,0.09,5.1,NULL,NULL,NULL,NULL,0.17,0.16,4.93,NULL),
+(347,6,'Carne, bovina, costela, assada',43.2,373,28.8,27.7,95,0,NULL,1,28,20,NULL,179,2.2,92,270,0.08,5.5,NULL,NULL,NULL,NULL,0.08,0.35,0.56,NULL),
+(348,6,'Carne, bovina, costela, crua',52.7,358,16.7,31.8,44,0,NULL,0.9,NULL,12,NULL,130,1.2,70,151,NULL,2.7,5,5,5,0.12,0.11,NULL,5.99,NULL),
+(349,6,'Carne, bovina, coxão duro, sem gordura, cozido',58.5,217,31.9,8.9,71,0,NULL,0.9,4,14,NULL,189,1.7,41,252,0.07,5,2,2,2,NULL,NULL,NULL,3.26,NULL),
+(350,6,'Carne, bovina, coxão duro, sem gordura, cru',69.8,148,21.5,6.2,60,0,NULL,1.1,3,21,NULL,189,1.9,49,358,0.05,2.8,2,2,2,0.12,0.2,NULL,4.38,NULL),
+(351,6,'Carne, bovina, coxão mole, sem gordura, cozido',58,219,32.4,8.9,84,0,NULL,1.2,4,13,NULL,183,2.6,44,239,0.11,4.7,2,2,2,NULL,NULL,NULL,4.09,NULL),
+(352,6,'Carne, bovina, coxão mole, sem gordura, cru',68.6,169,21.2,8.7,84,0,NULL,1,3,21,NULL,175,1.9,61,335,0.05,2.6,3,3,3,0.12,0.19,NULL,2.32,NULL),
+(353,6,'Carne, bovina, cupim, assado',48.4,330,28.6,23,91,0,NULL,1,8,18,NULL,212,2.7,72,321,0.08,5.3,NULL,NULL,NULL,NULL,0.08,0.14,2.16,NULL),
+(354,6,'Carne, bovina, cupim, cru',64.8,221,19.5,15.3,51,0,NULL,0.9,4,13,NULL,220,1.1,47,151,0.03,2.4,3,3,3,0.11,0.05,0.03,3.33,NULL),
+(355,6,'Carne, bovina, fígado, cru',71.3,141,20.7,5.4,393,1.1,NULL,1.5,4,12,0.26,334,5.6,76,265,9.01,3.5,7937,7937,7937,0.14,0.9,NULL,10.66,NULL),
+(356,6,'Carne, bovina, fígado, grelhado',55,225,29.9,9,601,4.2,NULL,2,6,10,0.22,420,5.8,82,309,12.58,4,14574,14574,14574,0.21,2.69,NULL,11.92,NULL),
+(357,6,'Carne, bovina, filé mingnon, sem gordura, cru',71.9,143,21.6,5.6,55,0,NULL,1.1,3,21,0.01,193,1.9,49,322,0.08,2.8,4,4,4,0.12,0.1,0.04,1.78,NULL),
+(358,6,'Carne, bovina, filé mingnon, sem gordura, grelhado',57,220,32.8,8.8,103,0,NULL,1.3,4,28,0.02,308,2.9,58,326,0.14,4.1,NULL,NULL,NULL,0.03,0.08,0.05,4.27,NULL),
+(359,6,'Carne, bovina, flanco, sem gordura, cozido',62,196,29.4,7.8,62,0,NULL,0.9,4,14,NULL,181,2.8,42,249,0.07,5.6,2,2,2,NULL,NULL,NULL,2.32,NULL),
+(360,6,'Carne, bovina, flanco, sem gordura, cru',72.1,141,20,6.2,50,0,NULL,1,3,18,NULL,167,1.6,54,324,0.05,4.5,2,2,2,0.11,0.16,NULL,3.28,NULL),
+(361,6,'Carne, bovina, fraldinha, com gordura, cozida',49.7,338,24.2,26,65,0,NULL,0.7,3,14,NULL,161,2,39,207,0.07,6.5,NULL,NULL,NULL,NULL,0.05,0.07,1.61,NULL),
+(362,6,'Carne, bovina, fraldinha, com gordura, crua',65.4,221,17.6,16.1,54,0,NULL,0.9,3,16,0,131,1.5,51,274,0.04,4.2,5,5,5,0.08,0.07,0.03,3.56,NULL),
+(363,6,'Carne, bovina, lagarto, cozido',57.6,222,32.9,9.1,56,0,NULL,0.9,4,13,NULL,167,1.9,48,254,0.05,7,3,3,3,NULL,NULL,NULL,2.55,NULL),
+(364,6,'Carne, bovina, lagarto, cru',71,135,20.5,5.2,56,0,NULL,1.1,3,20,NULL,185,1.3,54,362,0.05,2.4,2,2,2,0.1,0.14,0.03,2.19,NULL),
+(365,6,'Carne, bovina, língua, cozida',53.4,315,21.4,24.8,105,0,NULL,0.7,6,12,0.01,136,2.1,59,175,0.08,4.1,NULL,NULL,NULL,NULL,0.08,NULL,1.83,NULL),
+(366,6,'Carne, bovina, língua, crua',65,215,17.1,15.8,118,0,NULL,0.8,5,15,0.02,164,1.7,73,251,0.09,2.9,NULL,NULL,NULL,0.05,0.04,NULL,1.18,NULL),
+(367,6,'Carne, bovina, maminha, crua',70,153,20.9,7,51,0,NULL,1,3,16,NULL,181,1.1,37,274,0.03,3.5,3,3,3,0.12,0.08,0.03,4.27,NULL),
+(368,6,'Carne, bovina, maminha, grelhada',65.3,153,30.7,2.4,88,0,NULL,1.4,4,21,NULL,237,2.4,58,386,0.08,5.6,NULL,NULL,NULL,NULL,0.04,0.13,1.46,NULL),
+(369,6,'Carne, bovina, miolo de alcatra, sem gordura, cru',69.5,163,21.6,7.8,60,0,NULL,1,3,20,0.01,165,2,43,299,0.06,3,4,4,4,0.12,0.07,NULL,3.59,NULL),
+(370,6,'Carne, bovina, miolo de alcatra, sem gordura, grelhado',52.4,241,31.9,11.6,92,0,NULL,1.2,5,26,0.02,279,3.2,52,385,0.11,4.8,NULL,NULL,NULL,0.03,0.05,0.05,4.66,NULL),
+(371,6,'Carne, bovina, músculo, sem gordura, cozido',62.8,194,31.2,6.7,56,0,NULL,0.9,5,13,NULL,176,2.4,62,253,0.08,6.4,2,2,2,NULL,NULL,NULL,1.97,NULL),
+(372,6,'Carne, bovina, músculo, sem gordura, cru',72.4,142,21.6,5.5,51,0,NULL,1,4,17,NULL,162,1.9,66,296,0.05,3.7,2,2,2,0.09,0.19,0.04,2.79,NULL),
+(373,6,'Carne, bovina, paleta, com gordura, crua',70.6,159,21.4,7.5,58,0,NULL,0.9,4,14,NULL,158,1.8,65,250,0.08,3.7,NULL,NULL,NULL,NULL,0.04,NULL,0.91,NULL),
+(374,6,'Carne, bovina, paleta, sem gordura, cozida',62.9,194,29.7,7.4,56,0,NULL,0.8,6,18,NULL,197,2.2,58,250,0.1,6.8,NULL,NULL,NULL,0.03,0.04,NULL,2.16,NULL),
+(375,6,'Carne, bovina, paleta, sem gordura, crua',72.1,141,21,5.7,42,0,NULL,1,4,18,NULL,163,1.9,66,319,0.05,3.3,3,3,3,0.13,0.21,NULL,2.67,NULL),
+(376,6,'Carne, bovina, patinho, sem gordura, cru',72.9,133,21.7,4.5,56,0,NULL,1,3,20,0.01,170,1.8,49,318,0.05,4.5,2,2,2,0.11,0.08,NULL,3.61,NULL),
+(377,6,'Carne, bovina, patinho, sem gordura, grelhado',55.2,219,35.9,7.3,126,0,NULL,1.3,5,27,0.02,289,3,60,421,0.12,8.1,NULL,NULL,NULL,0.04,0.03,NULL,3.01,NULL),
+(378,6,'Carne, bovina, peito, sem gordura, cozido',51.2,338,22.2,27,100,0,NULL,0.7,4,14,NULL,136,1.6,56,204,0.05,3.9,NULL,NULL,NULL,NULL,0.3,NULL,4.54,NULL),
+(379,6,'Carne, bovina, peito, sem gordura, cru',61.5,259,17.6,20.4,59,0,NULL,0.9,4,15,NULL,124,1.3,64,241,0.06,2.6,4,4,4,0.11,0.07,0.03,3.91,NULL),
+(380,6,'Carne, bovina, picanha, com gordura, crua',65.6,213,18.8,14.7,60,0,NULL,0.9,2,14,NULL,165,1.7,38,232,0.06,3.8,3,3,3,0.13,0.16,NULL,4.77,NULL),
+(381,6,'Carne, bovina, picanha, com gordura, grelhada',53.4,289,26.4,19.5,92,0,NULL,1.1,4,24,NULL,246,3.2,60,355,0.14,5.5,NULL,NULL,NULL,0.03,0.06,0.13,1.6,NULL),
+(382,6,'Carne, bovina, picanha, sem gordura, crua',72.4,134,21.3,4.7,75,0,NULL,1,3,20,NULL,183,2.1,61,322,0.08,4.2,NULL,NULL,NULL,NULL,0.05,NULL,1.89,NULL),
+(383,6,'Carne, bovina, picanha, sem gordura, grelhada',54.6,238,31.9,11.3,100,0,NULL,1.2,4,25,0.02,282,3.6,61,377,0.17,6.7,NULL,NULL,NULL,NULL,0.08,0.03,1.82,NULL),
+(384,6,'Carne, bovina, seca, cozida',47.2,313,26.9,21.9,100,0,NULL,4.7,13,12,0.02,82,1.9,1943,86,0.03,7.7,NULL,NULL,NULL,NULL,0.06,NULL,2.32,NULL),
+(385,6,'Carne, bovina, seca, crua',39.2,313,19.7,25.4,92,0,NULL,15.3,14,12,0.01,100,1.3,4440,190,NULL,3.7,NULL,NULL,NULL,NULL,0.07,NULL,2.83,NULL),
+(386,6,'Coxinha de frango, frita',42.2,283,9.6,11.8,15,34.5,5,1.8,18,17,0.28,93,1.3,532,166,0.09,0.5,NULL,NULL,NULL,0.09,NULL,NULL,2.6,NULL),
+(387,6,'Croquete, de carne, cru',56,246,12,15.6,30,13.9,NULL,2.5,15,24,0.34,144,2.5,711,221,0.1,2.7,NULL,NULL,NULL,0.11,0.04,NULL,1.95,NULL),
+(388,6,'Croquete, de carne, frito',39.2,347,16.9,22.7,38,18.1,NULL,3.1,18,30,0.39,176,2.3,916,313,0.09,3.3,NULL,NULL,NULL,0.14,0.04,NULL,4.77,NULL),
+(389,6,'Empada de frango, pré-cozida, assada',28.2,358,6.9,15.6,23,47.5,2.2,1.8,16,18,0.32,78,1.2,525,138,0.07,0.6,NULL,NULL,NULL,0.12,NULL,NULL,NULL,NULL),
+(390,6,'Empada, de frango, pré-cozida',32,377,7.3,22.9,23,35.5,2.2,2.2,14,17,0.25,78,0.7,771,156,0.09,0.5,NULL,NULL,NULL,0.11,NULL,0.03,NULL,NULL),
+(391,6,'Frango, asa, com pele, crua',67.5,213,18.1,15.1,113,0,NULL,0.7,11,23,0.01,155,0.6,96,211,0.02,1.2,10,10,10,0.11,0.04,NULL,2.58,NULL),
+(392,6,'Frango, caipira, inteiro, com pele, cozido',59.7,243,23.9,15.6,110,0,NULL,0.8,17,18,0.01,162,1.7,56,210,0.08,1.7,16,16,16,0.03,NULL,NULL,6.33,NULL),
+(393,6,'Frango, caipira, inteiro, sem pele, cozido',61.4,196,29.6,7.7,106,0,NULL,1.1,66,23,0.02,210,2.1,53,224,0.15,2.7,6,6,6,0.03,0.03,NULL,9.2,NULL),
+(394,6,'Frango, coração, cru',69.1,222,12.6,18.6,159,0,NULL,0.8,6,20,0.05,193,4.1,95,220,0.2,2,9,9,9,0.23,0.35,NULL,3.39,NULL),
+(395,6,'Frango, coração, grelhado',63.5,207,22.4,12.1,280,0.6,NULL,1.3,8,20,0.06,276,6.5,128,243,0.3,3.4,NULL,NULL,NULL,0.2,0.19,NULL,9.7,NULL),
+(396,6,'Frango, coxa, com pele, assada',59.8,215,28.5,10.4,145,0.1,NULL,1.3,8,14,NULL,251,1.2,95,318,0.05,2.6,6,6,6,0.05,0.05,NULL,10.4,NULL),
+(397,6,'Frango, coxa, com pele, crua',72.9,161,17.1,9.8,97,0,NULL,0.8,8,26,0.02,185,0.7,95,275,0.03,2,10,10,10,0.16,0.05,NULL,2.04,NULL),
+(398,6,'Frango, coxa, sem pele, cozida',66.7,167,26.9,5.8,133,0,NULL,0.9,12,11,NULL,187,0.8,64,191,0.03,2.8,NULL,NULL,NULL,0.07,NULL,NULL,8.2,NULL),
+(399,6,'Frango, coxa, sem pele, crua',76.4,120,17.8,4.9,91,0,NULL,0.9,8,27,0.02,196,0.8,98,291,0.03,2.2,12,12,12,0.17,0.05,NULL,3.53,NULL),
+(400,6,'Frango, fígado, cru',77.8,106,17.6,3.5,341,0,NULL,1.2,6,28,0.35,344,9.5,82,281,0.26,3.7,3863,3863,3863,0.62,0.56,NULL,6.36,NULL),
+(401,6,'Frango, filé, à milanesa',54.9,221,28.5,7.8,84,7.5,1.1,1.3,9,35,0.06,249,1.1,122,408,0.05,0.8,7,7,7,0.05,0.04,0.52,11.07,NULL),
+(402,6,'Frango, inteiro, com pele, cru',66.5,226,16.4,17.3,85,0,NULL,0.7,6,24,0.01,174,0.6,63,217,0.04,1.1,7,7,7,0.08,0.03,NULL,3.28,NULL),
+(403,6,'Frango, inteiro, sem pele, assado',63.2,187,28,7.5,111,0,NULL,1.2,9,14,NULL,233,0.6,70,283,0.03,1.6,NULL,NULL,NULL,0.09,NULL,NULL,11.2,NULL),
+(404,6,'Frango, inteiro, sem pele, cozido',67.5,170,25,7.1,99,0,NULL,0.9,8,12,NULL,194,0.5,51,217,0.04,1.2,NULL,NULL,NULL,0.1,NULL,NULL,12.83,NULL),
+(405,6,'Frango, inteiro, sem pele, cru',74.9,129,20.6,4.6,78,0,NULL,0.9,7,27,0.01,190,0.5,73,238,0.03,1.2,4,4,4,0.12,0.03,NULL,3.28,NULL),
+(406,6,'Frango, peito, com pele, assado',58.5,212,33.4,7.6,109,0,NULL,1.5,8,18,0.01,297,0.5,56,380,0.01,1,6,6,6,0.12,NULL,NULL,15.8,NULL),
+(407,6,'Frango, peito, com pele, cru',71.9,149,20.8,6.7,80,0,NULL,0.9,8,28,0.01,213,0.4,62,252,0.05,0.6,4,4,4,0.09,NULL,NULL,2.58,NULL),
+(408,6,'Frango, peito, sem pele, cozido',65.6,163,31.5,3.2,89,0,NULL,1,6,14,NULL,224,0.3,36,231,0.02,0.9,NULL,NULL,NULL,0.1,NULL,NULL,7.6,NULL),
+(409,6,'Frango, peito, sem pele, cru',74.8,119,21.5,3,59,0,NULL,1,7,31,0.01,222,0.4,56,267,0.03,0.7,2,2,2,0.1,NULL,NULL,5.86,NULL),
+(410,6,'Frango, peito, sem pele, grelhado',63.8,159,32,2.5,89,0,NULL,1.4,5,18,NULL,295,0.3,50,387,0.02,0.8,NULL,NULL,NULL,0.11,NULL,NULL,24.83,NULL),
+(411,6,'Frango, sobrecoxa, com pele, assada',55,260,28.7,15.2,158,0,NULL,1.3,11,15,NULL,252,1.2,96,323,0.06,2.2,8,8,8,0.1,0.05,NULL,11.2,NULL),
+(412,6,'Frango, sobrecoxa, com pele, crua',63.6,255,15.5,20.9,88,0,NULL,0.8,7,22,0.01,154,0.7,68,190,0.05,1.3,7,7,7,0.09,0.06,NULL,3.4,NULL),
+(413,6,'Frango, sobrecoxa, sem pele, assada',55.6,233,29.2,12,145,0,NULL,1.5,12,17,NULL,281,1.2,106,382,0.07,2.2,11,11,11,0.1,0.05,NULL,10.2,NULL),
+(414,6,'Frango, sobrecoxa, sem pele, crua',72.7,162,17.6,9.6,84,0,NULL,0.9,6,26,0.02,187,0.9,80,241,0.06,1.7,4,4,4,0.12,0.06,NULL,4.06,NULL),
+(415,6,'Hambúrguer, bovino, cru',63.6,215,13.2,16.2,70,4.2,NULL,2.9,34,25,0.36,141,1.9,869,383,0.16,1.7,NULL,NULL,NULL,0.06,0.08,0.04,3.53,NULL),
+(416,6,'Hambúrguer, bovino, frito',52.5,258,20,17,49,6.3,NULL,4.2,62,60,0.52,324,3,1252,660,0.17,3.2,NULL,NULL,NULL,0.13,0.06,NULL,2.73,NULL),
+(417,6,'Hambúrguer, bovino, grelhado',59.2,210,13.2,12.4,59,11.3,NULL,3.9,56,48,0.4,263,2.6,1090,538,0.18,3,NULL,NULL,NULL,0.18,0.06,NULL,5.77,NULL),
+(418,6,'Lingüiça, frango, crua',64.8,218,14.2,17.4,64,0,NULL,3.3,11,19,0.05,182,0.5,1126,280,0.05,0.7,NULL,NULL,NULL,0.11,0.05,NULL,4.6,NULL),
+(419,6,'Lingüiça, frango, frita',59.6,245,18.3,18.5,76,0,NULL,3.8,15,29,0.1,262,0.8,1374,364,0.04,1.2,NULL,NULL,NULL,0.11,0.05,NULL,5.67,NULL),
+(420,6,'Lingüiça, frango, grelhada',58.6,244,18.2,18.4,80,0,NULL,3.8,14,21,0.1,228,0.7,1351,356,0.09,1,NULL,NULL,NULL,0.12,0.04,NULL,5.97,NULL),
+(421,6,'Lingüiça, porco, crua',62.5,227,16.1,17.6,53,0,NULL,3.2,6,14,0.01,157,0.4,1176,316,0.04,1.4,NULL,NULL,NULL,0.49,0.05,NULL,2.6,NULL),
+(422,6,'Lingüiça, porco, frita',54.6,280,20.5,21.3,75,0,NULL,3.9,8,18,0.01,211,0.9,1432,409,0.06,3.1,NULL,NULL,NULL,0.41,0.07,NULL,5.83,NULL),
+(423,6,'Lingüiça, porco, grelhada',50.5,296,23.2,21.9,82,0,NULL,4,8,19,0.01,210,1,1456,427,0.07,3.5,NULL,NULL,NULL,0.4,0.07,NULL,6.6,NULL),
+(424,6,'Mortadela',56.4,269,12,21.6,83,5.8,NULL,4.1,67,19,0.11,216,1.5,1212,247,0.08,1,25,NULL,NULL,0.14,0.06,NULL,3,NULL),
+(425,6,'Peru, congelado, assado',65.3,163,26.2,5.7,91,0,NULL,2.2,14,12,0.02,197,0.6,628,175,0.03,1.2,NULL,NULL,NULL,0.06,NULL,NULL,6.23,NULL),
+(426,6,'Peru, congelado, cru',78.2,94,18.1,1.8,68,0,NULL,2.5,10,19,NULL,217,0.9,711,281,0.36,1.4,NULL,NULL,NULL,0.06,NULL,NULL,4.23,NULL),
+(427,6,'Porco, bisteca, crua',67.7,164,21.5,8,56,0,NULL,1,6,24,NULL,195,0.5,54,335,0.07,1.4,NULL,NULL,NULL,0.9,NULL,NULL,2.65,NULL),
+(428,6,'Porco, bisteca, frita',47.3,311,33.7,18.5,126,0,NULL,1.6,69,29,0.01,290,0.8,63,404,0.07,2.2,10,10,10,0.34,NULL,NULL,3.25,NULL),
+(429,6,'Porco, bisteca, grelhada',51.8,280,28.9,17.4,82,0,NULL,1.2,34,25,NULL,229,0.9,51,366,0.06,2.3,NULL,NULL,NULL,0.77,0.14,0.03,1.92,NULL),
+(430,6,'Porco, costela, assada',36.9,402,30.2,30.3,113,0,NULL,1.4,17,14,NULL,201,1,63,246,0.07,3.1,NULL,NULL,NULL,0.71,0.05,NULL,10.63,NULL),
+(431,6,'Porco, costela, crua',61.2,256,18,19.8,69,0,NULL,0.9,15,18,NULL,159,0.9,88,248,0.05,2.3,NULL,NULL,NULL,0.62,NULL,NULL,8.27,NULL),
+(432,6,'Porco, lombo, assado',56.6,210,35.7,6.4,103,0,NULL,1.4,20,18,NULL,238,0.5,39,311,0.03,1.8,NULL,NULL,NULL,0.75,0.07,0.11,12.43,NULL),
+(433,6,'Porco, lombo, cru',67.7,176,22.6,8.8,55,0,NULL,1,4,24,NULL,195,0.5,53,334,0.01,0.9,NULL,NULL,NULL,0.95,NULL,NULL,13.83,NULL),
+(434,6,'Porco, orelha, salgada, crua',59.7,258,18.5,19.9,83,0,NULL,2,5,2,0.01,31,1.4,616,228,0.23,0.6,NULL,NULL,NULL,0.07,0.03,NULL,NULL,NULL),
+(435,6,'Porco, pernil, assado',49.3,262,32.1,13.9,110,0,NULL,1.4,18,27,0.01,247,1.3,62,395,0.09,3.3,NULL,NULL,NULL,0.77,0.09,NULL,6.57,NULL),
+(436,6,'Porco, pernil, cru',67.1,186,20.1,11.1,59,0,NULL,1,13,23,NULL,192,0.9,102,256,0.16,1.7,NULL,NULL,NULL,1.06,0.06,NULL,5.67,NULL),
+(437,6,'Porco, rabo, salgado, cru',43.8,377,15.6,34.5,89,0,NULL,3.5,22,4,0.01,42,0.6,1158,24,0.05,1.4,NULL,NULL,NULL,0.2,0.06,NULL,1.43,NULL),
+(438,6,'Presunto, com capa de gordura',73.9,128,14.4,6.8,40,1.4,NULL,3.5,12,17,0.04,274,0.7,1021,295,0.04,1.3,NULL,NULL,NULL,0.62,0.03,NULL,3.78,NULL),
+(439,6,'Presunto, sem capa de gordura',77.2,94,14.3,2.7,36,2.1,NULL,3.7,23,18,0.04,244,0.8,1039,307,0.04,1.5,NULL,NULL,NULL,0.5,0.03,NULL,4,NULL),
+(440,6,'Quibe, assado',69,136,14.6,2.7,34,12.9,1.9,0.9,16,36,0.76,174,2.2,40,288,0.51,4.1,NULL,NULL,NULL,0.04,0.04,NULL,1.81,NULL),
+(441,6,'Quibe, cru',74.5,109,12.4,1.7,27,10.8,1.6,0.7,12,26,0.39,126,1.7,39,242,0.13,2.8,NULL,NULL,NULL,0.09,0.03,NULL,1.2,NULL),
+(442,6,'Quibe, frito',54,254,14.9,15.8,38,12.3,NULL,3,22,39,0.72,166,2,836,322,0.16,2.8,NULL,NULL,NULL,0.12,0.1,NULL,1.5,NULL),
+(443,6,'Salame',34.7,398,25.8,30.6,85,2.9,NULL,5.9,87,30,0.03,354,1.3,1574,548,0.05,3.2,NULL,NULL,NULL,0.88,0.23,NULL,12.49,NULL),
+(444,6,'Toucinho, cru',27.6,593,11.5,60.3,73,0,NULL,0.3,2,4,NULL,35,0.4,50,58,0.11,0.2,NULL,NULL,NULL,NULL,NULL,NULL,5.1,NULL),
+(445,6,'Toucinho, frito',6.3,697,27.3,64.3,89,0,NULL,0.7,9,9,0.01,95,0.9,125,171,0.1,0.8,NULL,NULL,NULL,NULL,NULL,NULL,5.03,NULL),
+(446,7,'Bebida láctea, pêssego',87.7,55,2.1,1.9,5,7.6,0.3,0.7,89,9,NULL,63,NULL,46,62,0.02,0.2,NULL,NULL,NULL,0.03,0.14,NULL,0.33,2.1),
+(447,7,'Creme de Leite',70.9,221,1.5,22.5,66,4.5,NULL,0.6,83,8,0.01,118,0.3,52,119,0.02,0.3,128,NULL,NULL,NULL,0.1,NULL,0,NULL),
+(448,7,'Iogurte, natural',90,51,4.1,3,14,1.9,NULL,0.9,143,11,NULL,119,NULL,52,71,0.02,0.4,23,NULL,NULL,0.04,0.22,NULL,NULL,0.9),
+(449,7,'Iogurte, natural, desnatado',89.2,41,3.8,0.3,3,5.8,NULL,0.9,157,12,NULL,110,NULL,60,182,NULL,0.5,NULL,NULL,NULL,NULL,0.22,NULL,NULL,0.3),
+(450,7,'Iogurte, sabor abacaxi',NULL,NULL,NULL,NULL,6,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(451,7,'Iogurte, sabor morango',84.6,70,2.7,2.3,7,9.7,0.2,0.6,101,8,NULL,73,NULL,38,52,0.02,0.3,27,NULL,NULL,0.03,0.12,0.02,NULL,NULL),
+(452,7,'Iogurte, sabor pêssego',85.1,68,2.5,2.3,8,9.4,0.7,0.6,95,8,NULL,66,0.1,37,52,0.02,0.3,21,NULL,NULL,0.02,0.1,NULL,0.38,NULL),
+(453,7,'Leite, condensado',27,313,7.7,6.7,28,57,NULL,1.6,246,22,NULL,187,0.1,94,329,0.03,0.9,53,NULL,NULL,0.06,0.33,0.57,NULL,2.1),
+(454,7,'Leite, de cabra',87.1,66,3.1,3.8,14,5.2,NULL,0.9,112,10,NULL,113,0.1,74,140,0.04,0.4,35,NULL,NULL,NULL,0.14,NULL,NULL,NULL),
+(455,7,'Leite, de vaca, achocolatado',80.9,83,2.1,2.2,6,14.2,0.6,0.7,70,13,0.05,71,0.5,72,155,0.04,0.3,39,NULL,NULL,0.07,0.29,0.14,1.23,3.3),
+(456,7,'Leite, de vaca, desnatado, pó',3.1,362,34.7,0.9,25,53,NULL,8.2,1363,109,NULL,1673,0.9,432,1556,0.16,3.8,299,NULL,NULL,0.32,1.2,NULL,0.81,NULL),
+(457,7,'Leite, de vaca, desnatado, UHT',NULL,NULL,NULL,NULL,4,NULL,NULL,0.8,134,10,NULL,85,NULL,51,140,0.02,0.4,NULL,NULL,NULL,0.04,0.26,NULL,1.49,NULL),
+(458,7,'Leite, de vaca, integral',NULL,NULL,NULL,NULL,10,NULL,NULL,0.8,123,10,NULL,82,NULL,64,133,0.02,0.4,21,NULL,NULL,0.04,0.24,NULL,1.52,NULL),
+(459,7,'Leite, de vaca, integral, pó',2.7,497,25.4,26.9,85,39.2,NULL,5.8,890,77,NULL,1242,0.5,323,1132,0.11,2.7,361,NULL,NULL,0.24,1.03,NULL,NULL,NULL),
+(460,7,'Leite, fermentado',81.9,70,1.9,0.1,2,15.7,NULL,0.5,72,6,0.02,63,NULL,33,94,0.01,0.3,NULL,NULL,NULL,0.08,0.11,NULL,NULL,0.5),
+(461,7,'Queijo, minas, frescal',56.1,264,17.4,20.2,62,3.2,NULL,3,579,7,0.02,123,0.9,31,105,NULL,0.3,161,NULL,NULL,NULL,0.25,NULL,NULL,NULL),
+(462,7,'Queijo, minas, meia cura',47.1,321,21.2,24.6,76,3.6,NULL,3.5,696,27,0.02,403,0.2,501,120,0.07,2.7,111,NULL,NULL,0.03,0.32,NULL,NULL,NULL),
+(463,7,'Queijo, mozarela',45.3,330,22.6,25.2,80,3,NULL,3.8,875,24,0.03,470,0.3,581,62,0.08,3.5,109,NULL,NULL,NULL,0.2,NULL,NULL,NULL),
+(464,7,'Queijo, parmesão',21.2,453,35.6,33.5,106,1.7,NULL,8,992,33,0.05,745,0.5,1844,96,0.17,4.4,66,NULL,NULL,NULL,0.44,0.23,NULL,NULL),
+(465,7,'Queijo, pasteurizado',54.4,303,9.4,27.4,82,5.7,NULL,3.1,323,16,0.01,578,0.3,780,194,0.05,1.3,57,NULL,NULL,NULL,0.2,NULL,NULL,NULL),
+(466,7,'Queijo, petit suisse, morango',72.2,121,5.8,2.8,12,18.5,NULL,0.7,731,27,0.02,448,0.1,412,121,0.01,2.7,273,NULL,NULL,0.22,0.27,0.21,2.68,NULL),
+(467,7,'Queijo, prato',42.4,360,22.7,29.1,91,1.9,NULL,3.9,940,28,0.03,461,0.3,580,73,0.1,3.5,123,NULL,NULL,NULL,0.22,NULL,NULL,NULL),
+(468,7,'Queijo, requeijão, cremoso',62.5,257,9.6,23.4,74,2.4,NULL,2,259,12,0.02,448,0.1,558,93,0.05,1.3,195,NULL,NULL,NULL,0.19,NULL,NULL,NULL),
+(469,7,'Queijo, ricota',73.6,140,12.6,8.1,49,3.8,NULL,1.9,253,12,NULL,162,0.1,283,112,NULL,0.5,53,NULL,NULL,NULL,0.15,NULL,NULL,NULL),
+(470,8,'Bebida isotônica, sabores variados',93.5,26,0,0,NULL,6.4,NULL,0.1,1,NULL,NULL,9,0.7,44,13,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(471,8,'Café, infusão 10%',97.4,9,0.7,0.1,NULL,1.5,NULL,0.4,3,10,0.04,9,NULL,1,156,0.01,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(472,8,'Cana, aguardente 1',NULL,216,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,3,NULL,0.06,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(473,8,'Cana, caldo de',81.7,65,NULL,NULL,NULL,18.2,0.1,0.1,9,12,0.21,5,0.8,NULL,18,0.01,0.1,NULL,NULL,NULL,NULL,NULL,0.03,NULL,2.8),
+(474,8,'Cerveja, pilsen 2',92.4,41,0.6,NULL,NULL,3.3,NULL,0.1,5,7,0.01,19,NULL,4,29,NULL,NULL,NULL,NULL,NULL,NULL,NULL,0.14,2.92,NULL),
+(475,8,'Chá, erva-doce, infusão 5%',99.6,1,0,0,NULL,0.4,NULL,0,2,1,0.01,1,NULL,1,10,0.01,NULL,NULL,NULL,NULL,1.23,0.23,NULL,NULL,NULL),
+(476,8,'Chá, mate, infusão 5%',99.3,3,0,0.1,NULL,0.6,NULL,0,1,2,0.27,NULL,NULL,NULL,5,NULL,NULL,NULL,NULL,NULL,0.89,NULL,NULL,NULL,NULL),
+(477,8,'Chá, preto, infusão 5%',99.4,2,0,0,NULL,0.6,NULL,0,0,1,0.09,2,NULL,NULL,13,NULL,NULL,NULL,NULL,NULL,3.11,0.48,NULL,NULL,NULL),
+(478,8,'Coco, água de',94.3,22,0,0,NULL,5.3,0.1,0.5,19,5,0.25,4,NULL,2,162,NULL,NULL,NULL,NULL,NULL,0.01,NULL,NULL,NULL,2.4),
+(479,8,'Refrigerante, tipo água tônica',92,31,0,0,NULL,8,NULL,0,1,NULL,NULL,0,NULL,8,2,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(480,8,'Refrigerante, tipo cola',91.3,34,0,0,NULL,8.7,NULL,0.1,1,NULL,NULL,17,NULL,7,1,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(481,8,'Refrigerante, tipo guaraná',90,39,0,0,NULL,10,NULL,0,1,NULL,NULL,NULL,NULL,9,1,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(482,8,'Refrigerante, tipo laranja',88.2,46,0,0,NULL,11.8,NULL,0,2,1,NULL,1,NULL,9,16,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(483,8,'Refrigerante, tipo limão',89.7,40,0,0,NULL,10.3,NULL,0,2,1,NULL,NULL,NULL,9,4,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(484,9,'Omelete, de queijo',60.5,268,15.6,22,384,0.4,NULL,1.5,166,14,0.03,314,1.4,216,127,0.03,1.4,59,NULL,NULL,0.03,0.24,NULL,NULL,NULL),
+(485,9,'Ovo, de codorna, inteiro, cru',71.7,177,13.7,12.7,568,0.8,NULL,1.2,79,11,NULL,279,3.3,129,79,0.04,2.1,305,NULL,NULL,0.11,0.12,NULL,0.97,NULL),
+(486,9,'Ovo, de galinha, clara, cozida/10minutos',85.2,59,13.4,0.1,NULL,0,NULL,0.7,6,11,NULL,15,0.1,181,146,0.03,NULL,NULL,NULL,NULL,NULL,0.08,NULL,NULL,NULL),
+(487,9,'Ovo, de galinha, gema, cozida/10minutos',50,353,15.9,30.8,1272,1.6,NULL,1.7,114,9,0.06,386,2.9,45,87,NULL,2.9,148,NULL,NULL,0.18,0.22,NULL,NULL,NULL),
+(488,9,'Ovo, de galinha, inteiro, cozido/10minutos',75.8,146,13.3,9.5,397,0.6,NULL,0.8,49,11,0.02,184,1.5,146,139,0.04,1.2,32,NULL,NULL,0.08,0.3,NULL,NULL,NULL),
+(489,9,'Ovo, de galinha, inteiro, cru',75.6,143,13,8.9,356,1.6,NULL,0.8,42,13,NULL,164,1.6,168,150,0.06,1.1,79,NULL,NULL,0.07,0.58,NULL,0.75,NULL),
+(490,9,'Ovo, de galinha, inteiro, frito',63.5,240,15.6,18.6,516,1.2,NULL,1.1,73,16,0.03,422,2.1,166,184,0.04,1.5,94,NULL,NULL,0.06,0.32,NULL,NULL,NULL),
+(491,10,'Achocolatado, pó',1.1,401,4.2,2.2,NULL,91.2,3.9,1.4,44,77,0.55,200,5.4,65,496,0.56,1,796,NULL,NULL,1.38,1.02,1.52,4.99,NULL),
+(492,10,'Açúcar, cristal',0.1,387,0.3,NULL,NULL,99.6,NULL,NULL,8,1,NULL,NULL,0.2,NULL,3,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(493,10,'Açúcar, mascavo',3.3,369,0.8,0.1,NULL,94.5,NULL,1.4,127,80,2.03,38,8.3,25,522,0.17,0.5,NULL,NULL,NULL,NULL,0.03,NULL,NULL,NULL),
+(494,10,'Açúcar, refinado',0.1,387,0.3,NULL,NULL,99.5,NULL,NULL,4,1,NULL,NULL,0.1,12,6,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(495,10,'Chocolate, ao leite',1.3,540,7.2,30.3,17,59.6,2.2,1.7,191,57,0.3,212,1.6,77,355,0.31,1.1,NULL,NULL,NULL,0.05,0.22,0.59,0.63,NULL),
+(496,10,'Chocolate, ao leite, com castanha do Pará',1.2,559,7.4,34.2,16,55.4,2.5,1.8,171,80,0.36,303,1.5,64,431,0.45,1.3,36,NULL,NULL,0.04,0.24,NULL,1.37,1.4),
+(497,10,'Chocolate, ao leite, dietético',1.7,557,6.9,33.8,13,56.3,2.8,1.3,188,67,0.41,276,3.3,85,458,0.43,1.1,7,NULL,NULL,0.31,0.25,NULL,0.79,2),
+(498,10,'Chocolate, meio amargo',1,475,4.9,29.9,2,62.4,4.9,1.8,45,107,0.83,220,3.6,9,432,0.77,1.5,NULL,NULL,NULL,0.2,0.04,NULL,1.06,2.1),
+(499,10,'Cocada branca',3.4,449,1.1,13.6,NULL,81.4,3.6,0.5,7,17,0.36,396,1.2,29,183,0.2,0.4,NULL,NULL,NULL,NULL,NULL,0.16,0.47,NULL),
+(500,10,'Doce, de abóbora, cremoso',43.9,199,0.9,0.2,NULL,54.6,2.3,0.4,13,6,0.01,14,0.9,NULL,137,NULL,NULL,NULL,NULL,NULL,0.08,NULL,0.04,NULL,0.1),
+(501,10,'Doce, de leite, cremoso',27.5,306,5.5,6,20,59.5,NULL,1.5,195,16,0.01,141,0.1,120,259,0.02,0.5,36,36,36,0.05,0.26,NULL,NULL,NULL),
+(502,10,'Geléia, mocotó, natural',73.5,106,2.1,0.1,NULL,24.2,NULL,0.1,4,1,NULL,NULL,0.1,43,2,NULL,NULL,NULL,NULL,NULL,0.19,NULL,0.22,NULL,NULL),
+(503,10,'Glicose de milho',20.4,292,0,0,NULL,79.4,NULL,0.2,6,2,NULL,5,0.1,59,5,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(504,10,'Maria mole',21.6,301,3.8,0.2,NULL,73.6,0.7,0.9,13,7,0.06,4,0.4,15,24,0.2,0.1,NULL,NULL,NULL,0.04,NULL,NULL,NULL,NULL),
+(505,10,'Maria mole, coco queimado',20.7,307,3.9,0.1,NULL,75.1,0.6,0.2,19,6,0.09,5,0.5,14,36,0.03,0.1,NULL,NULL,NULL,0.04,NULL,NULL,NULL,NULL),
+(506,10,'Marmelada',28.5,257,0.4,0.1,NULL,70.8,4.1,0.2,11,6,0.06,20,0.7,11,83,0.09,0.1,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(507,10,'Mel, de abelha',15.8,309,0,0,NULL,84,NULL,0.1,10,6,0.38,4,0.3,6,99,NULL,0.2,NULL,NULL,NULL,0.11,NULL,NULL,NULL,0.7),
+(508,10,'Melado',22.1,297,0,0,NULL,76.6,NULL,1.3,102,115,2.62,74,5.4,4,395,0.84,0.3,NULL,NULL,NULL,NULL,0.05,0.2,NULL,NULL),
+(509,10,'Quindim',23.9,411,4.7,24.4,271,46.3,3.2,0.6,37,15,0.25,168,1.4,27,111,0.1,1.1,76,NULL,NULL,NULL,0.04,0.04,1.14,NULL),
+(510,10,'Rapadura',7.1,352,1,0.1,NULL,90.8,NULL,1.1,30,47,1.66,21,4.4,22,459,0.17,0.6,NULL,NULL,NULL,NULL,NULL,0.04,NULL,NULL),
+(511,11,'Café, pó, torrado',2.9,419,14.7,11.9,NULL,65.8,51.2,4.7,107,165,2.58,169,8.1,1,1609,1.3,0.5,NULL,NULL,NULL,NULL,NULL,NULL,11.89,NULL),
+(512,11,'Capuccino, pó',2.6,417,11.3,8.6,29,73.6,2.4,3.8,467,71,0.17,358,2.3,382,886,NULL,1.1,52,NULL,NULL,NULL,1.41,NULL,NULL,NULL),
+(513,11,'Fermento em pó, químico',7.1,90,0.5,0.1,NULL,43.9,NULL,48.5,NULL,NULL,NULL,NULL,NULL,10052,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(514,11,'Fermento, biológico, levedura, tablete',71.9,90,17,1.5,NULL,7.7,4.2,1.9,18,38,0.2,419,2.6,40,576,0.29,11,NULL,NULL,NULL,0.3,0.36,NULL,NULL,NULL),
+(515,11,'Gelatina, sabores variados, pó',1.2,380,8.9,NULL,NULL,89.2,NULL,0.6,27,2,0.03,2,0.3,235,7,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,40),
+(516,11,'Sal, dietético',0.6,NULL,NULL,NULL,NULL,NULL,NULL,99.4,NULL,NULL,NULL,NULL,NULL,23432,20468,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(517,11,'Sal, grosso',0.4,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,39943,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(518,11,'Shoyu',70.6,61,3.3,0.3,NULL,11.6,NULL,14.1,15,24,0.07,47,0.5,5024,165,NULL,0.2,NULL,NULL,NULL,0.83,0.06,0.15,2.1,NULL),
+(519,11,'Tempero a base de sal',7.7,21,2.7,0.3,NULL,2.1,0.6,87.3,NULL,NULL,NULL,NULL,NULL,32560,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(520,12,'Azeitona, preta, conserva',68.5,194,1.2,20.3,NULL,5.5,4.6,4.5,59,5,0.06,16,5.5,1567,79,0.25,0.3,NULL,NULL,NULL,NULL,NULL,0.04,NULL,NULL),
+(521,12,'Azeitona, verde, conserva',76.3,137,0.9,14.2,NULL,4.1,3.8,4.5,46,4,0.03,5,0.2,1347,20,0.14,0.1,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(522,12,'Chantilly, spray, com gordura vegetal',55.1,315,0.5,27.3,NULL,16.9,NULL,0.2,2,1,NULL,16,NULL,110,5,0.01,0.1,NULL,NULL,NULL,0.07,NULL,NULL,NULL,NULL),
+(523,12,'Leite, de coco',78,166,1,18.4,NULL,2.2,0.7,0.4,6,17,0.24,26,0.5,44,144,0.16,0.3,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(524,12,'Maionese, tradicional com ovos',58.4,302,0.6,30.5,42,7.9,NULL,2.6,3,1,NULL,14,0.1,787,16,NULL,0.1,8,NULL,NULL,NULL,0.05,NULL,NULL,NULL),
+(525,13,'Acarajé',50.5,289,8.3,19.9,25,19.1,9.4,2.1,124,51,0.59,142,1.9,305,354,0.23,1.2,NULL,NULL,NULL,0.06,NULL,NULL,NULL,NULL),
+(526,13,'Arroz carreteiro',68,154,10.8,7.1,36,11.6,1.5,2.4,13,9,0.18,48,1,1622,87,0.08,2.7,NULL,NULL,NULL,0.03,NULL,0.07,1.6,NULL),
+(527,13,'Baião de dois, arroz e feijão-de-corda',69.1,136,6.2,3.2,4,20.4,5.1,1.1,33,19,0.27,72,0.6,93,157,0.08,0.6,NULL,NULL,NULL,0.04,NULL,0.04,NULL,NULL),
+(528,13,'Barreado',71.2,165,18.3,9.5,60,0.2,0.1,0.8,15,21,0.05,169,2.4,48,295,0.17,4.8,NULL,NULL,NULL,0.03,0.04,NULL,2.09,NULL),
+(529,13,'Bife à cavalo, com contra filé',54.2,291,23.7,21.1,257,0,NULL,1.1,26,19,0.02,216,2.1,83,272,0.06,3.2,37,NULL,NULL,0.05,0.1,NULL,2.92,NULL),
+(530,13,'Bolinho de arroz',41.5,274,8,8.3,70,41.7,2.7,0.5,24,13,0.42,87,2.1,59,96,0.13,0.9,NULL,NULL,NULL,0.08,0.05,NULL,NULL,NULL),
+(531,13,'Camarão à baiana',82,101,7.9,6,117,3.2,0.4,0.9,43,15,0.11,244,1.4,85,139,0.07,0.5,NULL,NULL,NULL,NULL,NULL,NULL,0.31,4.1),
+(532,13,'Charuto, de repolho',81.5,78,6.8,1.1,21,10.1,1.5,0.4,23,13,0.22,68,0.9,12,184,1.45,1.8,NULL,NULL,NULL,0.03,NULL,NULL,0.32,4.8),
+(533,13,'Cuscuz, de milho, cozido com sal',71.1,113,2.2,0.7,NULL,25.3,2.1,0.7,2,3,0.02,23,0.2,248,11,NULL,0.2,NULL,NULL,NULL,NULL,NULL,0.06,NULL,NULL),
+(534,13,'Cuscuz, paulista',68.9,142,2.6,4.6,15,22.5,2.4,1.3,14,5,0.06,26,0.3,236,53,0.05,0.2,NULL,NULL,NULL,NULL,NULL,NULL,0.93,NULL),
+(535,13,'Cuxá, molho',81,80,5.6,3.6,58,5.7,3,4,105,34,0.23,111,0.9,1344,124,0.17,0.6,NULL,33,16,0.04,NULL,0.03,1.1,10.3),
+(536,13,'Dobradinha',75.3,125,19.8,4.4,144,0,NULL,0.3,11,8,0.11,57,1,29,58,0.06,2.7,NULL,NULL,NULL,NULL,NULL,NULL,4.27,NULL),
+(537,13,'Estrogonofe de carne',70.1,173,15,10.8,66,3,NULL,1.1,28,22,0.03,186,2.7,123,322,0.14,2,28,NULL,NULL,0.06,0.05,NULL,3.25,NULL),
+(538,13,'Estrogonofe de frango',70.9,157,17.6,8,80,2.6,NULL,1,26,25,0.03,195,1.5,99,307,0.12,0.6,28,NULL,NULL,0.07,0.04,0.03,3.37,NULL),
+(539,13,'Feijão tropeiro mineiro',61.7,152,10.2,6.8,68,19.6,3.6,1.8,41,36,0.38,199,2.2,365,349,0.22,1.4,9,NULL,NULL,NULL,NULL,NULL,1.44,NULL),
+(540,13,'L',71.8,117,8.7,6.5,22,11.6,5.1,1.4,32,32,0.24,105,1.3,278,303,0.15,0.8,NULL,NULL,NULL,0.08,0.03,NULL,0.93,NULL),
+(541,13,'Frango, com açafrão',79.5,113,9.7,6.2,50,4.1,0.2,0.5,13,16,0.09,167,0.8,29,256,0.02,0.5,8,NULL,NULL,0.03,NULL,NULL,1.64,5.3),
+(542,13,'Macarrão, molho bolognesa',71.4,120,4.9,0.9,7,22.5,0.8,0.2,11,10,0.22,54,1.4,9,84,0.09,0.8,NULL,NULL,NULL,0.04,NULL,NULL,1.44,NULL),
+(543,13,'Maniçoba',76.4,134,10,8.7,43,3.4,2.2,1.5,66,24,0.6,55,3.2,407,148,0.16,2,NULL,222,111,0.04,0.07,0.05,2.58,NULL),
+(544,13,'Quibebe',81.1,86,8.6,2.7,NULL,6.6,1.7,1,8,10,0.05,43,0.8,247,153,0.61,1.6,NULL,105,52,NULL,0.03,0.04,NULL,NULL),
+(545,13,'Salada, de legumes, com maionese',82,96,1.1,7,7,8.9,2.2,0.9,12,9,0.1,22,0.2,228,141,0.04,0.2,NULL,283,141,0.04,NULL,0.04,NULL,NULL),
+(546,13,'Salada, de legumes, cozida no vapor',90,35,2,0.3,NULL,7.1,2.5,0.6,33,19,0.24,45,0.4,3,244,0.17,0.3,NULL,510,255,0.05,0.03,0.03,NULL,29.4),
+(547,13,'Salpicão, de frango',72.5,148,13.9,7.8,53,4.6,0.4,1.1,9,13,0.03,103,0.3,248,149,0.08,0.4,NULL,15,7,0.05,NULL,NULL,1.48,9.3),
+(548,13,'Sarapatel',75,123,18.5,4.4,315,1.1,NULL,1.1,12,13,0.1,164,7.2,216,199,1.48,1.8,1464,NULL,NULL,NULL,0.11,NULL,3.84,NULL),
+(549,13,'Tabule',85.8,57,2,1.2,NULL,10.6,2.1,0.4,19,18,0.37,35,0.6,1,188,0.12,0.6,NULL,NULL,NULL,0.03,NULL,NULL,0.6,16.2),
+(550,13,'Tacacá',85.2,47,7,0.4,71,3.4,0.2,4.1,45,30,0.12,89,0.9,1349,240,0.22,0.8,NULL,35,18,0.05,NULL,NULL,NULL,NULL),
+(551,13,'Tapioca, com manteiga',24.9,348,0.1,10.9,31,63.6,NULL,0.5,30,3,0.06,8,0.2,158,19,0.01,0,75,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(552,13,'Tucupi, com pimenta-de-cheiro',91.9,27,2.1,0.3,NULL,4.7,0.2,1,28,42,0.52,31,1.1,5,391,0.12,0.9,NULL,51,25,NULL,NULL,0.05,NULL,46.4),
+(553,13,'Vaca atolada',75,145,5.1,9.3,19,10.1,2.3,0.5,63,16,0.04,72,0.7,26,220,0.12,1.2,NULL,NULL,NULL,NULL,NULL,NULL,0.77,NULL),
+(554,13,'Vatapá',58.4,255,6,23.2,44,9.7,1.7,2.7,47,39,0.48,108,1.4,880,209,0.28,0.9,NULL,NULL,NULL,NULL,NULL,NULL,1.8,NULL),
+(555,13,'Virado à paulista',48.6,307,10.2,25.6,66,14.1,2.2,1.5,41,22,0.15,135,1.1,346,237,0.07,1,10,NULL,NULL,0.14,0.05,NULL,2.6,NULL),
+(556,13,'Yakisoba',69.4,113,7.5,2.6,NULL,18.3,1.1,2.2,14,13,0.15,83,0.6,794,159,0.06,0.7,NULL,62,31,0.16,NULL,NULL,2.09,NULL),
+(557,14,'Amendoim, grão, cru',6.4,544,27.2,43.9,NULL,20.3,8,2.2,NULL,171,1.96,407,2.5,NULL,580,0.78,3.2,NULL,NULL,NULL,0.1,0.03,0.76,10.18,NULL),
+(558,14,'Amendoim, torrado, salgado',1.7,606,22.5,54,NULL,18.7,7.8,3.2,39,159,1.7,261,1.3,376,496,0.68,2.1,NULL,NULL,NULL,0.08,NULL,0.21,7.47,NULL),
+(559,14,'Ervilha, em vagem',76.8,88,7.5,0.5,NULL,14.2,9.7,1,24,42,0.4,152,1.4,NULL,311,0.2,1.2,NULL,NULL,NULL,0.27,0.07,0.06,1.16,12.4),
+(560,14,'Ervilha, enlatada, drenada',80.1,74,4.6,0.4,NULL,13.4,5.1,1.4,22,23,0.46,79,1.4,372,147,0.14,0.9,NULL,9,5,0.07,0.03,NULL,NULL,NULL),
+(561,14,'Feijão, carioca, cozido',80.4,76,4.8,0.5,NULL,13.6,8.5,0.7,27,42,0.28,87,1.3,2,255,0.19,0.7,NULL,NULL,NULL,0.04,NULL,NULL,NULL,NULL),
+(562,14,'Feijão, carioca, cru',14,329,20,1.3,NULL,61.2,18.4,3.5,123,210,1.02,385,8,NULL,1352,0.79,2.9,NULL,NULL,NULL,0.17,NULL,0.65,4.02,NULL),
+(563,14,'Feijão, fradinho, cozido',80,78,5.1,0.6,NULL,13.5,7.5,0.8,17,38,0.53,85,1.1,1,253,0.1,1.1,NULL,NULL,NULL,0.12,NULL,NULL,NULL,NULL),
+(564,14,'Feijão, fradinho, cru',12.7,339,20.2,2.4,NULL,61.2,23.6,3.5,78,178,1.43,355,5.1,10,1083,0.7,3.9,NULL,NULL,NULL,0.14,0.03,0.26,NULL,NULL),
+(565,14,'Feijão, jalo, cozido',75.8,93,6.1,0.5,NULL,16.5,13.9,1,29,44,0.32,121,1.9,1,348,0.24,1,NULL,NULL,NULL,0.13,NULL,0.04,NULL,NULL),
+(566,14,'Feijão, jalo, cru',13.5,328,20.1,0.9,NULL,61.5,30.3,3.9,98,170,0.99,427,7,25,1276,0.95,3,NULL,NULL,NULL,0.1,NULL,0.12,NULL,NULL),
+(567,14,'Feijão, preto, cozido',80.2,77,4.5,0.5,NULL,14,8.4,0.8,29,40,0.37,88,1.5,2,256,0.2,0.7,NULL,NULL,NULL,0.06,NULL,0.03,NULL,NULL),
+(568,14,'Feijão, preto, cru',14.9,324,21.3,1.2,NULL,58.8,21.8,3.8,111,188,1.32,471,6.5,NULL,1416,0.83,2.9,NULL,NULL,NULL,0.12,NULL,0.59,4.6,NULL),
+(569,14,'Feijão, rajado, cozido',77.9,85,5.5,0.4,NULL,15.3,9.3,0.9,29,42,0.29,113,1.4,1,315,0.23,0.9,NULL,NULL,NULL,0.09,NULL,0.04,NULL,NULL),
+(570,14,'Feijão, rajado, cru',15,326,17.3,1.2,NULL,62.9,24,3.7,111,170,1.17,335,18.6,14,1135,0.84,2.6,NULL,NULL,NULL,0.07,0.03,0.07,NULL,NULL),
+(571,14,'Feijão, rosinha, cozido',82.6,68,4.5,0.5,NULL,11.8,4.8,0.6,19,43,0.46,90,1.2,2,241,0.09,1.3,NULL,NULL,NULL,NULL,NULL,NULL,3.69,NULL),
+(572,14,'Feijão, rosinha, cru',12,337,20.9,1.3,NULL,62.2,20.6,3.6,68,184,1.08,394,5.3,24,1109,0.6,4,NULL,NULL,NULL,0.16,0.03,NULL,3.92,NULL),
+(573,14,'Feijão, roxo, cozido',80,77,5.7,0.5,NULL,12.9,11.5,0.8,23,34,0.32,106,1.4,1,268,0.22,1,NULL,NULL,NULL,0.15,NULL,0.03,NULL,NULL),
+(574,14,'Feijão, roxo, cru',12.6,331,22.2,1.2,NULL,60,33.8,4,120,162,1.34,394,6.9,10,1221,1.04,3.3,NULL,NULL,NULL,0.26,NULL,0.08,NULL,NULL),
+(575,14,'Grão-de-bico, cru',12.3,355,21.2,5.4,NULL,57.9,12.4,3.2,114,146,3.16,342,5.4,5,1116,0.67,3.2,NULL,NULL,NULL,0.52,NULL,0.75,NULL,NULL),
+(576,14,'Guandu, cru',11.4,344,19,2.1,NULL,64,21.3,3.5,129,166,1.02,269,1.9,2,1215,0.57,2,NULL,NULL,NULL,1.06,NULL,0.07,2.69,1.5),
+(577,14,'Lentilha, cozida',76.3,93,6.3,0.5,NULL,16.3,7.9,0.6,16,22,0.29,104,1.5,1,220,0.17,1.1,NULL,NULL,NULL,0.03,NULL,NULL,NULL,NULL),
+(578,14,'Lentilha, crua',11.5,339,23.2,0.8,NULL,62,16.9,2.6,54,94,1.08,368,7,NULL,887,0.83,3.5,NULL,NULL,NULL,0.11,NULL,0.42,5.07,NULL),
+(579,14,'Paçoca, amendoim',1.8,487,16,26.1,NULL,52.4,7.3,3.8,22,101,1.06,198,1.1,167,348,0.38,1.6,NULL,NULL,NULL,NULL,NULL,1.2,NULL,NULL),
+(580,14,'Pé-de-moleque, amendoim',2.9,503,13.2,28,NULL,54.7,3.4,1.1,27,108,1.09,171,1.3,16,355,0.43,1.4,NULL,NULL,NULL,0.04,0.03,0.11,5.94,NULL),
+(581,14,'Soja, farinha',5.8,404,36,14.6,NULL,38.4,20.2,5.1,206,242,2.87,539,13.1,6,1922,1.29,4.5,NULL,NULL,NULL,0.2,0.04,0.03,NULL,NULL),
+(582,14,'Soja, extrato solúvel, natural, fluido',91.3,39,2.4,1.6,NULL,4.3,0.4,0.5,17,15,0.15,53,0.4,57,121,0.08,0.3,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(583,14,'Soja, extrato solúvel, pó',4.5,459,35.7,26.2,NULL,28.5,7.3,5.2,359,216,2.68,647,7,83,1607,1.19,5.8,NULL,NULL,NULL,NULL,0.11,0.35,NULL,9.2),
+(584,14,'Soja, queijo (tofu)',86.6,64,6.6,4,NULL,2.1,0.8,0.7,81,38,0.33,130,1.4,1,182,0.18,0.9,NULL,NULL,NULL,0.04,NULL,0.03,NULL,NULL),
+(585,14,'Tremoço, cru',9.7,381,33.6,10.3,NULL,43.8,32.3,2.6,177,121,NULL,265,2.8,3,708,0.79,4.2,NULL,NULL,NULL,0.24,NULL,0.07,NULL,25),
+(586,14,'Tremoço, em conserva',67.7,121,11.1,3.8,NULL,12.4,14.4,5,16,4,NULL,40,0.3,1809,5,0.27,0.6,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(587,15,'Amêndoa, torrada, salgada',3.1,581,18.6,47.3,NULL,29.5,11.6,1.5,237,222,1.95,493,3.1,279,640,0.93,2.6,NULL,NULL,NULL,0.29,0.16,NULL,NULL,NULL),
+(588,15,'Castanha-de-caju, torrada, salgada',3.5,570,18.5,46.3,NULL,29.1,3.7,2.6,33,237,1.59,594,5.2,125,671,1.92,4.7,NULL,NULL,NULL,0.29,0.05,0.39,NULL,NULL),
+(589,15,'Castanha-do-Brasil, crua',3.5,643,14.5,63.5,NULL,15.1,7.9,3.4,146,365,1.1,853,2.3,1,651,1.79,4.2,NULL,NULL,NULL,0.3,NULL,0.44,NULL,NULL),
+(590,15,'Coco, cru',43,406,3.7,42,NULL,10.4,5.4,1,6,51,1,118,1.8,15,354,0.45,0.9,NULL,NULL,NULL,NULL,NULL,0.03,NULL,2.5),
+(591,15,'Coco,  verde, cru',NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL),
+(592,15,'Farinha, de mesocarpo de babaçu, crua',15.8,329,1.4,0.2,NULL,79.2,17.9,3.4,61,39,0.38,26,18.3,12,362,0.22,0.3,NULL,NULL,NULL,NULL,NULL,NULL,2.58,NULL),
+(593,15,'Gergelim, semente',3.9,584,21.2,50.4,NULL,21.6,11.9,2.9,825,361,2.67,741,5.4,3,546,1.51,5.2,NULL,NULL,NULL,0.94,NULL,0.13,5.92,NULL),
+(594,15,'Linhaça, semente',6.7,495,14.1,32.3,NULL,43.3,33.5,3.7,211,347,2.81,615,4.7,9,869,1.09,4.4,NULL,NULL,NULL,0.12,NULL,0.13,NULL,NULL),
+(595,15,'Pinhão, cozido',50.5,174,3,0.7,NULL,43.9,15.6,1.8,16,53,0.41,166,0.8,1,727,0.18,0.8,NULL,NULL,NULL,NULL,NULL,NULL,NULL,27.7),
+(596,15,'Pupunha, cozida',54.5,219,2.5,12.8,NULL,29.6,4.3,0.7,28,25,0.13,49,0.5,1,303,0.28,0.3,NULL,875,438,NULL,0.09,0.03,NULL,2.2),
+(597,15,'Noz, crua',6.2,620,14,59.4,NULL,18.4,7.2,2.1,105,153,4.05,396,2,5,533,0.75,2.1,NULL,NULL,NULL,0.38,NULL,0.13,1.08,NULL);
+
+    END;
+    $$;
+    `);
+    console.log("Comandos SQL submetidos ao SGBD.");
+  } catch (e: any) {
+    console.error("Erro ao submeter comandos SQL:", e.message);
+  } finally {
+    console.log("Finalizado");
+    pool.end();
+  }
+}
+
+load();
