--- server/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "server",
-  "version": "1.0.0",
-  "main": "src/index.ts",
-  "scripts": {
-    "create": "ts-node ./src/database/create",
-    "load": "ts-node ./src/database/load",
-    "start": "ts-node ./src",
-    "dev": "ts-node-dev ./src"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "description": "",
-  "dependencies": {
-    "bcrypt": "^5.1.1",
-    "cors": "^2.8.5",
-    "dotenv": "^16.4.5",
-    "express": "^4.19.2",
-    "jsonwebtoken": "^9.0.2",
-    "pg": "^8.12.0"
-  },
-  "devDependencies": {
-    "@types/bcrypt": "^5.0.2",
-    "@types/cors": "^2.8.17",
-    "@types/express": "^4.17.21",
-    "@types/jsonwebtoken": "^9.0.6",
-    "@types/pg": "^8.11.6",
-    "ts-node": "^10.9.2",
-    "ts-node-dev": "^2.0.0",
-    "typescript": "^5.4.5"
-  }
-}
