--- server/src/controllers/index.ts
@@ -1,6 +1,0 @@
-
-import UserController from "./UserController";
-
-export {
-    UserController,
-};
